% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsa.R
\name{bsa}
\alias{bsa}
\title{Body Surface Area Calculation}
\usage{
bsa(
  weight,
  height,
  weight_units = "kg",
  height_units = "m",
  method = "Mosteller",
  ...
)
}
\arguments{
\item{weight}{Body Weight (kg)}

\item{height}{Body Height (m)}

\item{weight_units}{If `weight` is in Units other than kg, specifiy here to  properly convert for calucations ("kg", "g", or "lbs")}

\item{height_units}{If `height` is in Units other than m, specifiy here to  properly convert for calucations ("m", "cm", or "in")}

\item{method}{Which BSA formula to use for calculation ()}
}
\value{
a numeric vector
}
\description{

}
\examples{
bsa_methods <- c("Mosteller", "DuBois", "Boyd", "Gehan-George", "Haycock", "Yu", "Livingston", "Tikuisis")
names(bsa_methods) <- bsa_methods
sapply(bsa_methods, \(x)bsa(weight = 70, height = 1.778, method = x))
bsa(weight = 70, height = 1.778)
bsa(weight = 70, height = 1.778, method = "Mosteller")
bsa(weight = 70, height = 1.778, method = "DuBois")
bsa(weight = 70, height = 1.778, method = "Boyd")
bsa(weight = 70, height = 1.778, method = "Gehan-George")
bsa(weight = 70, height = 1.778, method = "Haycock")
bsa(weight = 70, height = 1.778, method = "Yu")
bsa(weight = 70, height = 1.778, method = "Livingston")
bsa(weight = 70, height = 1.778, method = "Tikuisis")

# different units
bsa(weight = 154, weight_units = "lbs", 
    height = 70, height_units = "in", 
    method = "Mosteller")
bsa(weight = 70, height = 1.778)
}
