% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_age.R
\name{calculate_age}
\alias{calculate_age}
\title{Calculate age from date of birth}
\usage{
calculate_age(dob, date = "today", date_format = "ymd", units = "years")
}
\arguments{
\item{dob}{Date of birth, preferably "YYYY-MM-DD" format}

\item{date}{Date to calculate age from, default is today}

\item{date_format}{defaults to "ymd" "YYYY-MM-DD" format}

\item{units}{units to return age in, default is "years"; other options are "days", "weeks", "months"}
}
\value{
age in years
}
\description{
Calculate age from date of birth
}
\examples{
calculate_age(dob = "Jan 5, 1960", date_format = "mdy")
calculate_age(dob = "1/5/1960", date_format = "mdy")
calculate_age(dob = "1960-01-05", date_format = "ymd")
calculate_age(dob = "1960-01-05", date_format = "ymd", units = "days")
calculate_age(dob = "1960-01-05", date_format = "ymd", units = "weeks")
}
