% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ascvd_risk.R
\name{calculate_ascvd_risk}
\alias{calculate_ascvd_risk}
\title{Calculate 10-year ASCVD Risk (2014 original formula)}
\usage{
calculate_ascvd_risk(
  age,
  sex,
  race,
  total_chol,
  hdl_chol,
  sbp,
  using_antihypertensive_medication = FALSE,
  current_smoker = FALSE,
  diabetes = FALSE,
  chol_units = "mg/dL"
)
}
\arguments{
\item{age}{Age in years}

\item{sex}{Sex assigned at birth (Male/Female)}

\item{race}{Race self-identified}

\item{total_chol}{Total Cholesterol, mg/dL}

\item{hdl_chol}{HDL Cholesterol, mg/dL}

\item{sbp}{Systolic Blood Pressure, mmHg}

\item{using_antihypertensive_medication}{Hypertension medication use (TRUE/FALSE)}

\item{current_smoker}{Currently Smoking  (TRUE/FALSE)}

\item{diabetes}{Diabetes diganosis  (TRUE/FALSE)}

\item{chol_units}{Cholesterol units (default "mg/dL")}
}
\value{
A single value for 10-year Cardiovascular Risk. Multiple *100 to obtain as a percentage.
}
\description{
Calculate 10-year ASCVD Risk (2014 original formula)
}
\examples{
calculate_ascvd_risk(sex = "female",
                     race = "white",
                     age= 55,
                     total_chol = 213,
                     hdl_chol = 50,
                     sbp=120,
                     using_antihypertensive_medication=FALSE,
                     current_smoker=FALSE,
                     diabetes=FALSE)
n = 10
tibble(id = 1:n,
       sex = sample(c("female","male"), n, replace = T),
       race = sample(c("white","black"), n, replace = T),
       age= rnorm(n, 55),
       total_chol = rnorm(n, 213),
       hdl_chol = rnorm(n, 50),
       sbp=rnorm(n, 120),
       using_antihypertensive_medication=sample(c(TRUE, FALSE), n, replace = T),
       current_smoker=sample(c(TRUE, FALSE), n, replace = T),
       diabetes=sample(c(TRUE, FALSE), n, replace = T)) |> 
  mutate(ascvd_r = calculate_ascvd_risk_V(age = age, sex = sex, race=race, 
                                          total_chol = total_chol, hdl_chol=hdl_chol,
                                          sbp=sbp, using_antihypertensive_medication=using_antihypertensive_medication,
                                          current_smoker=current_smoker,
                                          diabetes=diabetes))
}
