% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_avg_glucose.R
\name{calculate_avg_glucose}
\alias{calculate_avg_glucose}
\title{Estimate average glucose from A1C value}
\usage{
calculate_avg_glucose(a1c, glucose_units = "mg/dl")
}
\arguments{
\item{a1c}{measured A1C%}

\item{glucose_units}{Desired glucose units for result- default is "mg/dl",
can indicate here for unit conversion options, eg "mg/dl" or "mmol/l"}
}
\value{
estimated average glucose with mg/dl default, as a single numeric value
}
\description{
Estimate average glucose from A1C value
}
\details{
Conversion from A1C to estimated average glucose using linear
interpolation described in the paper
\href{https://pubmed.ncbi.nlm.nih.gov/18540046/}{"Translating the A1C assay
into estimated average glucose values"}
}
\examples{
calculate_avg_glucose(5)
calculate_avg_glucose(5:12)
calculate_avg_glucose(5:12, glucose_units = "mM")
}
