% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bigtt.R
\name{calculate_bigtt}
\alias{calculate_bigtt}
\title{OGTT BIGTT estimates of beta cell function and insulin sensitivity}
\usage{
calculate_bigtt(
  time,
  glucose,
  insulin,
  gender = NA,
  bmi = NA,
  time_units = "min",
  glucose_units = "mg/dl",
  insulin_units = "uU/ml"
)
}
\arguments{
\item{time}{a column name (unquoted) indicating time values (in minutes)}

\item{glucose}{a column name (unquoted) storing glucose values (in mg/dL)}

\item{insulin}{a column name (unquoted) storing insulin values (in uU/mL)}

\item{gender}{Gender (M/F)}

\item{bmi}{body mass index (\eqn{kg/m^2})}

\item{time_units}{if units are not in "min", can indicate here for unit
conversion (options "min" or "hr")}

\item{glucose_units}{if units are not in "mg/dl", can indicate here for unit
conversion (options "mg/dl" or "mmol/l")}

\item{insulin_units}{if units are not in "uU/ml", can indicate here for unit
conversion (options "uU/ml" or "pmol/l")}
}
\value{
Dataframe of BIGTT Si and AIR estimates
}
\description{
This function calculates the BIGTT estimates of insulin sensitivity and beta
cell function using glucose and insulin sampled during a standard 75g oral
glucose tolerance test. Also requires BMI and gender.
}
\details{
Two estimates of SI are provided using either 0,30,120 or 0,60,120 values.
The calculations are described in
\href{https://pubmed.ncbi.nlm.nih.gov/17259491/}{Hansen et al}. The "full"
version is not included here due to using non-standard time points that I
have not collected.

Standard timepoints are 0, 30, 60, 90, and 120 min. Note: insulin unit
conversion may differ differ depending on assay. Insulin (pmol/l) = insulin
(uU/ml)*6

`calculate_bigtt()` accepts 3 separate vectors for time, glucose,
insulin and values for gender and BMI.
}
\examples{
time=c(0, 30, 60, 90, 120)              # minutes
glucose=c(93, 129, 178, 164, 97)        # mg/dL
insulin=c(12.8, 30.7, 68.5, 74.1, 44.0) # uU/mL
calculate_bigtt(time, glucose, insulin, gender="female", bmi=30)
calculate_bigtt(time, glucose, insulin, gender="Male", bmi=30) 
calculate_bigtt(time, glucose, insulin, gender="Male") 
calculate_bigtt(time, glucose, insulin, bmi=30) 
}
