% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bmi.R
\name{calculate_bmi}
\alias{calculate_bmi}
\title{Calculate Body Mass Index (BMI)}
\usage{
calculate_bmi(height, weight, height_units = "m", weight_units = "kg")
}
\arguments{
\item{height}{height in cm, m, or inches}

\item{weight}{weight in lbs or kg}

\item{height_units}{units in "cm", "m", or "inches"}

\item{weight_units}{units as "lbs" or "kg"}
}
\value{
numeric vector containing BMI results
}
\description{
Uses weight (kg) and height (m) to calculate BMI. Converts from lbs and inches if needed and units are properly supplied.
}
\examples{
calculate_bmi(height = 1.778, weight = 75)

calculate_bmi(height = 70, weight = 165,
              height_units = "inches", weight_units = "lbs")
              
#incorrect units
calculate_bmi(height = 70, weight = 165,
              height_units = "in", weight_units = "lbs")
              
calculate_bmi(height = c(70, 72, 65), 
              weight = c(165, 180, 155),
              height_units = "inches", weight_units = "lbs")
}
