% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cederholmisi.R
\name{calculate_cederholm_isi}
\alias{calculate_cederholm_isi}
\title{Cederholm Index Calculation (from standard 75g OGTT)}
\usage{
calculate_cederholm_isi(
  time,
  glucose,
  insulin,
  weight,
  time_units = "min",
  glucose_units = "mg/dl",
  insulin_units = "uU/ml",
  weight_units = "kg"
)
}
\arguments{
\item{time}{a column name (unquoted) indicating time values (in minutes)}

\item{glucose}{a column name (unquoted) storing glucose values (in mg/dL)}

\item{insulin}{a column name (unquoted) storing insulin values (in uU/mL)}

\item{time_units}{if units are not in "min", can indicate here for unit conversion (options "min" or "hr")}

\item{glucose_units}{if units are not in "mg/dl", can indicate here for unit conversion (options "mg/dl" or "mmol/l")}

\item{insulin_units}{if units are not in "uU/ml", can indicate here for unit conversion (options "uU/ml" or "pmol/l")}
}
\value{
Insulinogenic index as a single value
}
\description{
This function calculates the Cederholm Index using glucose and insulin 
sampled during a standard 75g oral glucose tolerance test.
}
\details{
Standard timepoints are 0, 30, 60, 90, and 120 min.
Note: insulin unit conversion may differ differ depending on assay. 
Insulin (pmol/l) = insulin (uU/ml)*6 

The formula used for this calculation is described in \href{https://pubmed.ncbi.nlm.nih.gov/2261853/}{Cederholm et al.}:
\deqn{\frac{75000 + (Glucose_{0} - Glucose_{120}) \cdot 1.15 \cdot 180 \cdot 0.19 \cdot Weight(kg)} 
{120 \cdot log(Insulin_{mean}) \cdot Glucose_{mean}}
}
}
\examples{
# individual objects for each item
time=c(0, 30, 60, 90, 120)              # minutes
glucose=c(93, 129, 178, 164, 97)        # mg/dL
insulin=c(12.8, 30.7, 68.5, 74.1, 44.0) # uU/mL
calculate_cederholm_isi(time, glucose, insulin, weight=70) # 11.80

# handling data stored in a dataframe
ogtt1 <- data.frame(time=c(0, 30, 60, 90, 120),              # minutes
                    glucose=c(93, 129, 178, 164, 97),        # mg/dL
                    insulin=c(12.8, 30.7, 68.5, 74.1, 44.0)) # uU/mL
calculate_cederholm_isi(ogtt1$time, ogtt1$glucose, ogtt1$insulin, weight=70) # 11.79529



# Convert units
ogtt5 <- data.frame(time = c(0,0.5,1,1.5,2), # time in hours
                    glucose = c(5.167, 7.167, 9.889, 9.111, 5.3889), # glucose in mmol/l
                    insulin = c(76.8,184.2,411,444.6,264)) # insulin in pmol/l

calculate_cederholm_isi(time =  ogtt5$time,
                        glucose = ogtt5$glucose,
                        insulin = ogtt5$insulin,
                        weight=154, weight_units="lbs",
                        time_units = "hr", insulin_units = "pmol/l", glucose_units = "mmol/l")
}
