% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_creatclearance_urine.R
\name{calculate_creatclearance_urine}
\alias{calculate_creatclearance_urine}
\title{Calculate 24hour creatinine clearance by 24-hour urine results}
\usage{
calculate_creatclearance_urine(
  urine_volume,
  creatinine_urine,
  creatinine_serum,
  urine_collection_time = 24,
  bsa_adjust = FALSE,
  height = NULL,
  weight = NULL
)
}
\arguments{
\item{weight}{Weight, in kg}

\item{age}{Age, in years}

\item{sex}{Sex}

\item{creatinine}{Serum creatinine, in mg/dL}
}
\value{
a numeric vector with eGFR (ml/min/1.73m^2)
}
\description{
Calculates Creatinine Clearance by 24hr urine collection results.
CrCL typically overestimates GFR by 10-20\% due to tubular secretion of creatinine. Other assumptions also apply including:
\itemize{
\item Creatinine/renal function is in steady state
\item Creatinine is filtered and not secreted (not true- leads to 10-20\% overestimate of GFR)
\item Complete and well-timed urine collection
}
}
\examples{
calculate_creatclearance_urine(urine_volume = 1000, creatinine_serum = 1.0, creatinine_urine = 100) # 69.4
calculate_creatclearance_urine(urine_volume = 1000, creatinine_serum = 1.0, creatinine_urine = 100,
                                   bsa_adjust = T) # error- need height, weight
calculate_creatclearance_urine(urine_volume = 1000, creatinine_serum = 1.0, creatinine_urine = 100,
                                   bsa_adjust = T, weight = 70, height = 1.5) # 70.3
}
