% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ddd.R
\name{calculate_ddd}
\alias{calculate_ddd}
\title{Calculating Defined Daily Dose (DDD) for hypertension medications}
\usage{
calculate_ddd(medication, dose)
}
\arguments{
\item{medication}{Medication name (internally converted to lowercase)}

\item{dose}{Dose of medication (in mg, total administered per day)}
}
\value{
calculated DDD value for the medication and dose provided
}
\description{
Calculation function for DDD. Provides values only for hypertension related
medications. The dose is the total administered per day, regardless of
frequency of administration. Medication name includes only the name of the
active ingredient, not the brand name (e.g. "clonidine" not "clonidine
hydrochloride"). The function will internally convert to lowercase for
matching purposes to the DDD list.
}
\examples{
# example data
htn_med_dose
htn_med_dose |>
  mutate(
    ddd_value = lookup_ddd(medication), # example only - not required
    ddd_calc = calculate_ddd(medication, dose)) 
}
