% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ecv.R
\name{calculate_ecv}
\alias{calculate_ecv}
\title{Estimate Extracellular Volume}
\usage{
calculate_ecv(
  age,
  height,
  weight,
  sex,
  weight_units = "kg",
  height_units = "cm",
  method = "Faucon"
)
}
\arguments{
\item{age}{Age in years}

\item{height}{Height (cm)}

\item{weight}{Weight (kg)}

\item{sex}{Sex (M/F)}

\item{weight_units}{Weight units if not in cm}

\item{height_units}{Height units if not in kg}

\item{method}{ECV Formula (default =  "Faucon"; also can use "Granerus")}
}
\value{
single value of ECV estimate in mL
}
\description{
Calculates estimated extracellular volume based on available population-based
regression equations (age, height, weight, sex). Formula described in
\href{https://pubmed.ncbi.nlm.nih.gov/35497800/}{Faucon et al. KI Rep. 2022}.
Comparison of different formulas suggests that the new formula provides more
accurate results. The "Granerus" equations have been used previously in
calculating single-sample iothalamate GFR and so this method is also
available.
}
\examples{
calculate_ecv(age=85, sex = "m", height = 187, weight=85)
calculate_ecv(age=85, sex = "m", height = 187, weight=85, method="Granerus")
calculate_ecv(age=85, sex = "f", height = 187, weight=85, method="Granerus")
n=5
df_test <-
  tibble(age=rnorm(n,50),
         height_cm=rnorm(n,167),
         height_m=rnorm(n,1.67),
         weight_kg=rnorm(n,80),
         sex=c(sample(c("M","F"), n, replace = T))
  )
df_test |> 
  mutate(calculate_ecv(age=age, 
                       height=height_cm,
                       weight=weight_kg,
                       sex=sex))
}
