% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_egfr_cg.R
\name{calculate_egfr_cg}
\alias{calculate_egfr_cg}
\title{Calculate estimated GFR by Cockcroft-Gault}
\usage{
calculate_egfr_cg(age, sex, weight, creatinine)
}
\arguments{
\item{age}{Age, in years}

\item{sex}{Sex}

\item{weight}{Weight, in kg}

\item{creatinine}{Serum creatinine, in mg/dL}
}
\value{
a numeric vector with eGFR (ml/min/1.73m^2)
}
\description{
Calculates estimated glomerular filtration rate (GFR) by several
  common methods. Converts weight to Kg and Height to cm if needed. Available
  methods for eGFR calculations include:

  *  CKD-EPI(2021)
  *  CKD-EPI Cystatin-C
  *  MDRD
  *  Cockcroft-Gault
}
\examples{
calculate_egfr_cg(age=70, sex="Male", weight=70, creatinine=1.0) # 95.2
calculate_egfr_cg(age=70, sex="Female", weight=70, creatinine=1.0) # 95.2
calculate_egfr_cg(age=50, sex="Male", weight=70, creatinine=1.0) # 95.2
calculate_egfr_cg(age=50, sex="Female", weight=70, creatinine=1.0) # 95.2
}
