% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_egfr.R, R/calculate_egfr_ckdepi.R
\name{calculate_egfr_ckdepi}
\alias{calculate_egfr_ckdepi}
\title{Calculate estimated GFR by CKD-EPI equation}
\usage{
calculate_egfr_ckdepi(age, sex, creatinine, race = NULL, version = "2021")

calculate_egfr_ckdepi(age, sex, creatinine, race = NULL, version = "2021")
}
\arguments{
\item{age}{Age, in years}

\item{sex}{Sex}

\item{creatinine}{Serum Creatinine, mg/dL}

\item{race}{Race limited to either "White" or "Black"}

\item{version}{Equation version, either "2021" or "original"}
}
\value{
a numeric vector with eGFR (ml/min/1.73m^2)

a numeric vector with eGFR (ml/min/1.73m^2)
}
\description{
Calculates estimated glomerular filtration rate (GFR) by several
  common methods. Converts weight to Kg and Height to cm if needed. Avaialbe
  methods for eGFR calculations include:

  *  CKD-EPI(2021): 142*(Cr/A)^B * 0.9938^age * (1.012 if Female)
  *  CKD-EPI Cystatin-C: 135 x (Scr/A)B x (Scys/C)D x 0.9961age x (0.963 if female)
  *  MDRD:
  *  Cockcroft-Gault:

Calculates estimated glomerular filtration rate (GFR) by several
  common methods. Converts weight to Kg and Height to cm if needed. Available
  methods for eGFR calculations include:

  *  CKD-EPI AS Equation (2021) incorporating only Age, Sex 
  *  CKD-EPI ASR Equation (original, 2009) incorporating Age, Sex, Race
  *  CKD-EPI Cystatin-C (not incorporated)
  *  MDRD  
  *  Cockcroft-Gault 

Online calculators can be used to compare reults for [CKD-EPI AS (2021)](https://www.kidney.org/professionals/kdoqi/gfr_calculator),
[CKD-EPI ASR](https://www.niddk.nih.gov/health-information/professionals/clinical-tools-patient-management/kidney-disease/laboratory-evaluation/glomerular-filtration-rate-calculators/historical)
}
\examples{
calculate_egfr(age=70, sex="Male", creatinine=0.8, method="CKD-EPI") # 95.2
calculate_egfr(age=50, sex="Female", creatinine=1.0, method="CKD-EPI") # 68.6  
# CKD-EPI 2021 version (new race-free creatinine-based equation)
# https://www.kidney.org/professionals/kdoqi/gfr_calculator
calculate_egfr_ckdepi(age=50, sex="Male", creatinine=0.6) # 95.2
calculate_egfr_ckdepi(age=70, sex="Male", creatinine=0.8, version = "2021") # 95.2
calculate_egfr_ckdepi(age=70, sex="Female", creatinine=0.8, version = "2021") # 95.2

# use of race-incorporated version not recommended (ASR);
# if you want to use race, use version = "original".
calculate_egfr_ckdepi(age=70, sex="Female", creatinine=0.8, race="White", version = "2021") # error
calculate_egfr_ckdepi(age=70, sex="Female", creatinine=0.8, race="White", version = "original") # 74

# original CKD-EPI equation; see [NIH eGFR calculator](https://www.niddk.nih.gov/health-information/professionals/clinical-tools-patient-management/kidney-disease/laboratory-evaluation/glomerular-filtration-rate-calculators/historical)
calculate_egfr_ckdepi(age=70, sex="Male", creatinine=0.8, version = "original", race = "Black") # 104.2
calculate_egfr_ckdepi(age=70, sex="Male", creatinine=0.8, version = "original", race = "White") # 89.9

# CKD-EPI AS (2021) RESULTS; Race-free equation
# Note: do not use race in function call; results same regardless of race
# results verified against Table S11 in Inker NEJM 2021: eGFR (AS) new
calculate_egfr_ckdepi(age=50, sex="Male", creatinine=0.6) # 118
calculate_egfr_ckdepi(age=50, sex="Male", creatinine=1)   # 92
calculate_egfr_ckdepi(age=50, sex="Male", creatinine=1.5) # 56
calculate_egfr_ckdepi(age=50, sex="Male", creatinine=2.0) # 40

calculate_egfr_ckdepi(age=75, sex="Male", creatinine=0.6) # 101
calculate_egfr_ckdepi(age=75, sex="Male", creatinine=1)   # 78
calculate_egfr_ckdepi(age=75, sex="Male", creatinine=1.5) # 48
calculate_egfr_ckdepi(age=75, sex="Male", creatinine=2.0) # 34

calculate_egfr_ckdepi(age=50, sex="Female", creatinine=0.6) # 109
calculate_egfr_ckdepi(age=50, sex="Female", creatinine=1)   # 69
calculate_egfr_ckdepi(age=50, sex="Female", creatinine=1.5) # 42
calculate_egfr_ckdepi(age=50, sex="Female", creatinine=2.0) # 30

calculate_egfr_ckdepi(age=75, sex="Female", creatinine=0.6) # 94
calculate_egfr_ckdepi(age=75, sex="Female", creatinine=1)   # 59
calculate_egfr_ckdepi(age=75, sex="Female", creatinine=1.5) # 36
calculate_egfr_ckdepi(age=75, sex="Female", creatinine=2.0) # 26


# CKD-EPI ASR (2021) RESULTS; Race-based equation
# this equation is equivalent to the original 2009 CKD-EPI equation
# The CKD-EPI ASR-NB (2021) = "Non-Black" is equivalent to using race="White"
# version = "original" allows specification of race = "White" or "Black"
# results verified against Table S11 in Inker NEJM 2021: eGFR (ASR) current
calculate_egfr_ckdepi(age=50, sex="Male", race="Black", creatinine=0.6, version = "original") # 135 (not 136)
calculate_egfr_ckdepi(age=50, sex="Male", race="Black", creatinine=1, version = "original")   # 101
calculate_egfr_ckdepi(age=50, sex="Male", race="Black", creatinine=1.5, version = "original") # 62
calculate_egfr_ckdepi(age=50, sex="Male", race="Black", creatinine=2.0, version = "original") # 44

calculate_egfr_ckdepi(age=75, sex="Male", race="Black", creatinine=0.6, version = "original") # 113 (not 114)
calculate_egfr_ckdepi(age=75, sex="Male", race="Black", creatinine=1, version = "original")   # 84 (not 85)
calculate_egfr_ckdepi(age=75, sex="Male", race="Black", creatinine=1.5, version = "original") # 52
calculate_egfr_ckdepi(age=75, sex="Male", race="Black", creatinine=2.0, version = "original") # 36 (not 37)

calculate_egfr_ckdepi(age=50, sex="Female", race="Black", creatinine=0.6, version = "original") # 123 (not 124)
calculate_egfr_ckdepi(age=50, sex="Female", race="Black", creatinine=1, version = "original")   # 76
calculate_egfr_ckdepi(age=50, sex="Female", race="Black", creatinine=1.5, version = "original") # 46 (not 47)
calculate_egfr_ckdepi(age=50, sex="Female", race="Black", creatinine=2.0, version = "original") # 33

calculate_egfr_ckdepi(age=75, sex="Female", race="Black", creatinine=0.6, version = "original") # 103 (not 104)
calculate_egfr_ckdepi(age=75, sex="Female", race="Black", creatinine=1, version = "original")   # 63 (not 64)
calculate_egfr_ckdepi(age=75, sex="Female", race="Black", creatinine=1.5, version = "original") # 39
calculate_egfr_ckdepi(age=75, sex="Female", race="Black", creatinine=2.0, version = "original") # 27 (not 28)

# race = "White" forces use of the CKD-EPI ASR-NB result
calculate_egfr_ckdepi(age=50, sex="Male", race="White", creatinine=0.6, version = "original") # 117
calculate_egfr_ckdepi(age=50, sex="Male", race="White", creatinine=1, version = "original")   # 87
calculate_egfr_ckdepi(age=50, sex="Male", race="White", creatinine=1.5, version = "original") # 53 (not 54)
calculate_egfr_ckdepi(age=50, sex="Male", race="White", creatinine=2.0, version = "original") # 38

calculate_egfr_ckdepi(age=75, sex="Male", race="White", creatinine=0.6, version = "original") # 98
calculate_egfr_ckdepi(age=75, sex="Male", race="White", creatinine=1, version = "original")   # 73
calculate_egfr_ckdepi(age=75, sex="Male", race="White", creatinine=1.5, version = "original") # 45
calculate_egfr_ckdepi(age=75, sex="Male", race="White", creatinine=2.0, version = "original") # 32

calculate_egfr_ckdepi(age=50, sex="Female", race="White", creatinine=0.6, version = "original") # 106 (not 107)
calculate_egfr_ckdepi(age=50, sex="Female", race="White", creatinine=1, version = "original")   # 65
calculate_egfr_ckdepi(age=50, sex="Female", race="White", creatinine=1.5, version = "original") # 40
calculate_egfr_ckdepi(age=50, sex="Female", race="White", creatinine=2.0, version = "original") # 29 (not 28)

calculate_egfr_ckdepi(age=75, sex="Female", race="White", creatinine=0.6, version = "original") # 88 (not 89)
calculate_egfr_ckdepi(age=75, sex="Female", race="White", creatinine=1, version = "original")   # 55
calculate_egfr_ckdepi(age=75, sex="Female", race="White", creatinine=1.5, version = "original") # 34
calculate_egfr_ckdepi(age=75, sex="Female", race="White", creatinine=2.0, version = "original") # 24
}
