% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_egfr_ckdepi_cysc.R
\name{calculate_egfr_ckdepi_cysc}
\alias{calculate_egfr_ckdepi_cysc}
\title{Calculate estimated GFR by 2021 CKD-EPI creatine/Cystatin C equation}
\usage{
calculate_egfr_ckdepi_cysc(age, sex, creatinine, cystatin)
}
\arguments{
\item{age}{Age, in years}

\item{sex}{Sex}

\item{creatinine}{creatinine (mg/dL)}

\item{cystatin}{cystatin C (mg/L)}
}
\value{
a numeric vector with eGFR (ml/min/1.73m^2)
}
\description{
Calculates estimated glomerular filtration rate (GFR) by several
  common methods. Converts weight to Kg and Height to cm if needed. Available
  methods for eGFR calculations include:

  *  CKD-EPI(2021)
  *  CKD-EPI Cystatin-C
  *  MDRD
  *  Cockcroft-Gault
}
\examples{
calculate_egfr_ckdepi_cysc(age=50, sex = "Female", creatinine=1.0, cystatin=1.5) # 54
calculate_egfr_ckdepi_cysc(age=50, sex = "Male", creatinine=1.0, cystatin=1.5) # 54
calculate_egfr_ckdepi_cysc(age=50, sex = "Female", creatinine=0.6, cystatin=0.7) # 115
calculate_egfr_ckdepi_cysc(age=50, sex = "Male", creatinine=0.6, cystatin=0.7) # 123
calculate_egfr_ckdepi_cysc(age=50, sex = "Female", creatinine=0.6, cystatin=0.8) # 111
calculate_egfr_ckdepi_cysc(age=50, sex = "Male", creatinine=0.6, cystatin=0.8) # 118
}
