% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_egfr_ekfc.R
\name{calculate_egfr_ekfc_nonv}
\alias{calculate_egfr_ekfc_nonv}
\title{Calculate estimated GFR, European Kidney Function Consortium (EKFC) 2021 formula}
\usage{
calculate_egfr_ekfc_nonv(creatinine, age, sex, creatinine_units = "mg/dL")
}
\arguments{
\item{creatinine}{Serum Creatinine, mg/dL}

\item{age}{Age, in years}

\item{sex}{Sex}

\item{creatinine_units}{Creatinine units, if different than mg/dL}
}
\value{
a numeric vector with eGFR (ml/min/1.73m^2)
}
\description{
Calculates the Race-free kidney-transplant-specific estimated
  glomerular filtration rate. Methods described in
  \href{https://pubmed.ncbi.nlm.nih.gov/33166224/}{Pottel et al, Annals Int Med. 2021}.
  An online calculator is found \href{https://ekfccalculator.pages.dev/?value=50&value=1&value=#results}{here}.
}
\examples{
# CKD-EPI 2021 version (new race-free creatinine-based equation)
# https://www.kidney.org/professionals/kdoqi/gfr_calculator
calculate_egfr_ckdepi(age=50, sex="Male", creatinine=0.6) # 95.2
calculate_egfr_ckdepi(age=70, sex="Male", creatinine=0.8, version = "2021") # 95.2
calculate_egfr_ckdepi(age=70, sex="Female", creatinine=0.8, version = "2021") # 95.2

# KRS 2023 version : transplant specific
calculate_egfr_krs(age=50, sex="Male", creatinine=1)
calculate_egfr_krs(age=50, sex="Female", creatinine=1)
calculate_egfr_krs(age=75, sex="Male", creatinine=1)
calculate_egfr_krs(age=75, sex="Female", creatinine=1)

# EKFC 2021: 
calculate_egfr_ekfc(1.2, age=50, sex="Male")
calculate_egfr_ekfc(1.2, age=19, sex="Male")
calculate_egfr_ekfc(1.2, age=15, sex="Male")

calculate_egfr_ekfc(1.2, age=50, sex="Female")
calculate_egfr_ekfc(1.2, age=19, sex="Female")
calculate_egfr_ekfc(1.2, age=15, sex="Female")
calculate_egfr_ekfc(1.2, age=10, sex="Female")
calculate_egfr_ekfc(1.2, age=5, sex="Female")

# original has been vectorized to handle multiple results
data.frame(age=c(5,10,15,35, 50), 
sex=rep("Male",5),
creatinine=rep(1.2,5)) |> 
dplyr::mutate(egfr=calculate_egfr_ekfc(creatinine, age, sex))
}
