% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_egfr_krs.R
\name{calculate_egfr_krs}
\alias{calculate_egfr_krs}
\title{Calculate estimated GFR, Kidney transplant Recipient Specific (KRS)}
\usage{
calculate_egfr_krs(age, sex, creatinine)
}
\arguments{
\item{age}{Age, in years}

\item{sex}{Sex}

\item{creatinine}{Serum Creatinine, mg/dL}

\item{version}{Equation version, either "2021" or "original"}
}
\value{
a numeric vector with eGFR (ml/min/1.73m^2)
}
\description{
Calculates the Race-free kidney-transplant-specific estimated
  glomerular filtration rate. Methods described in
  \href{https://pubmed.ncbi.nlm.nih.gov/37257905/}{Raynaud et al. BMJ. 2023}.
  An online calculator is found at
  \href{https://transplant-prediction-system.shinyapps.io/eGFR_equation_KTX/}
}
\examples{
# CKD-EPI 2021 version (new race-free creatinine-based equation)
# https://www.kidney.org/professionals/kdoqi/gfr_calculator
calculate_egfr_ckdepi(age=50, sex="Male", creatinine=0.6) # 95.2
calculate_egfr_ckdepi(age=70, sex="Male", creatinine=0.8, version = "2021") # 95.2
calculate_egfr_ckdepi(age=70, sex="Female", creatinine=0.8, version = "2021") # 95.2

# KRS 2023 version : transplant specific
calculate_egfr_krs(age=50, sex="Male", creatinine=1)
calculate_egfr_krs(age=50, sex="Female", creatinine=1)
calculate_egfr_krs(age=75, sex="Male", creatinine=1)
calculate_egfr_krs(age=75, sex="Female", creatinine=1)
}
