% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gomez.R
\name{calculate_gomez}
\alias{calculate_gomez}
\title{Gomez formulas for renal afferent/efferent resistance}
\usage{
calculate_gomez(GFR, RPF, Hct, TP, sbp, dbp, hr, DM = F)
}
\arguments{
\item{GFR}{Glomerular filtration rate measured by iohexol clearance or other
precise method (mL/min)}

\item{RPF}{Renal plasms flow, determined by PAH infusion (mL/min)}

\item{Hct}{Hematocrit (fraction)}

\item{TP}{Total protein concentration (mg/dL, plasma)}

\item{sbp}{Systolic Blood pressure (mmHg)}

\item{dbp}{Diastolic Blood pressure (mmHg)}

\item{hr}{Heart rate (bpm)}

\item{DM}{Diabetes (\code{T}/\code{F}). defaults to \code{FALSE}}
}
\value{
Data.frame of calculated values
}
\description{
Calculates renal resistance measures using the formulas originally reported
by \href{https://pubmed.ncbi.nlm.nih.gov/14888692/}{Gomez}, referred to as
the "Gomez Formulas". For review of this method see
\href{https://pubmed.ncbi.nlm.nih.gov/27605583/}{Bjornstad}. For discussion
of the additionally reported A value see \href{https://pubmed.ncbi.nlm.nih.gov/1860182/}{Kimura.}
}
\details{
Assumptions imposed by Gomez equations are as follows:
\enumerate{
\item Intrarenal vascular resistances are divided into afferent, postglomerular, and efferent
\item Hydrostatic pressures within the renal tubules, venules, Bowman’s space, and interstitium (PBow) are in equilibrium of 10 mmHg
\item Glomerulus is  in filtration disequilibrium
\item Gross filtration coefficient KF is 0.0867mL/s/ mmHg given a normal kidney.
}
}
\examples{
calculate_gomez(GFR=85/60, RPF=377/60, Hct=0.432,
                sbp=125, dbp=73, hr=70, #MAP=93,
                TP=6.8, DM=F)
}
