% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_h2fpef.R
\name{calculate_h2fpef}
\alias{calculate_h2fpef}
\title{H2PEF calculator
Returns the H2FPEF}
\usage{
calculate_h2fpef(bmi, htn_meds, afib, pa_pressure, age, filling_pressure)
}
\arguments{
\item{bmi}{BMI, kg/m^2}

\item{htn_meds}{Number of Hypertension medications}

\item{afib}{Atrial Fibrillation (T/F)}

\item{pa_pressure}{estimated Pulmonary Artery Pressure (mm/hg)}

\item{age}{Age (years)}

\item{filling_pressure}{(Doppler echo E/e')}
}
\description{
H2PEF calculator
Returns the H2FPEF
}
\examples{
calculate_h2fpef(bmi=31, htn_meds = 3, afib = F, pa_pressure = 15, age = F, filling_pressure = 6)
library(dplyr, purrr)
dat=data.frame(bmi=c(20, 31, 50),
               htn_meds=c(0,3,3),
               afib=c(F,F,T),
               pa_pressure=c(15, 35, 36),
               age=c(45, 45, 65),
               filling_pressure=c(6, 9, 10))

purrr::pmap_df(dat, calculate_h2fpef)
dat |> 
  mutate(purrr::pmap_df(dat, calculate_h2fpef))
}
