% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ibw_devine.R
\name{calculate_ibw_devine}
\alias{calculate_ibw_devine}
\title{Ideal Body Weight Calculation}
\usage{
calculate_ibw_devine(
  height,
  gender = NA,
  weight_units = "kg",
  height_units = "in",
  method = "Devine"
)
}
\description{
This version is only the Devine method, and does not use the `switch()`
function. This did not avoid the vectorization issue, however, and it does
not work on dataframes without converting to a vectorized version. Kept only for documentation, since it is no faster than the other version.
}
