% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_iga_risk_postbiopsy.R,
%   R/calculate_iga_risk_postbiopsy_norace.R
\name{calculate_iga_risk_postbiopsy}
\alias{calculate_iga_risk_postbiopsy}
\alias{calculate_iga_risk_postbiopsy_norace}
\title{International IgAN Prediction Tool post-biopsy}
\usage{
calculate_iga_risk_postbiopsy(
  egfr,
  sbp,
  dbp,
  proteinuria,
  age,
  mest_M,
  mest_E,
  mest_S,
  mest_T,
  time = 60,
  immunosuppression = FALSE,
  RASB = TRUE,
  race = "Caucasian",
  explanation = FALSE
)

calculate_iga_risk_postbiopsy_norace(
  egfr,
  sbp,
  dbp,
  proteinuria,
  age,
  mest_M,
  mest_E,
  mest_S,
  mest_T,
  time = 60,
  immunosuppression = FALSE,
  RASB = TRUE,
  explanation = FALSE
)
}
\arguments{
\item{egfr}{estimated GFR (ml/min/1.73m2)}

\item{sbp}{Systolic BP (mmHg)}

\item{dbp}{Diastolic BP (mmHg)}

\item{proteinuria}{Proteinuria (g/day)}

\item{age}{Age (years)}

\item{mest_M}{MEST score, Mesangial (M) hypercellularity (0,1)}

\item{mest_E}{MEST score, Endocapillary (E) hypercellularity (0,1)}

\item{mest_S}{MEST score, Segmental Sclerosis (0,1)}

\item{mest_T}{MEST score, interstitial fibrosis/tubular atrophy (0,1,2)}

\item{time}{Time frame for Risk estimate (months; defaults to 5 years = 60
months)}

\item{immunosuppression}{Immunosuppression use (TRUE/FALSE)}

\item{RASB}{Renin-angiotensin-aldosterone system blockade use (TRUE/FALSE)}

\item{race}{Race as `"Chinese", "Japanese", "Caucasian", or "Other"`}

\item{explanation}{Whether to output results as text description (TRUE/FALSE;
defaults to FALSE)}
}
\value{
a single value of risk estimate for kidney outcome (>50% decline in
 eGFR or ESRD) in the desired time interval
}
\description{
Post-Biopsy Risk estimator for Kidney outcomes in patients with IgA
Nephropathy, using the Oxford IgaN MEST Histology Score, estimated at a
time-point 1 year after biopsy using the formulas in the manuscript by
\href{https://pubmed.ncbi.nlm.nih.gov/35490842/}{Barbour et al, Kidney
International 2022}. The calculations have been validated against an available
online calculator at
\href{https://qxmd.com/calculate/calculator_839/international-igan-prediction-tool-post-biopsy-adults#}{Calculate
by QxMD}. Note if Race does not fit into the categories of "Chinese", "Japanese", or "Caucasian" then it is recommended to use the race-free equation.

`calculate_iga_risk_postbiopsy()` is the full Race-based model
`calculate_iga_risk_postbiopsy_norace()` Provides an estimate when race is not
known or if race is not well matched by Caucasian, Chinese, or Japanese
race/ethnicity
}
\examples{


# Online calcresult = 51.42\% vs 51.41738\%
calculate_iga_risk_postbiopsy(egfr = 50, sbp=150, dbp=90,  proteinuria=1.2, age=30,
                              race="Chinese", RASB=T,
                              mest_M=1,mest_E=1,mest_S=1,mest_T=1, time=60)

# Get text description of estimate
calculate_iga_risk_postbiopsy(egfr = 50, sbp=150, dbp=90,  proteinuria=1.2, age=30,
                              race="Chinese", RASB=T,
                              mest_M=1,mest_E=1,mest_S=1,mest_T=1, time=60,
                              explanation = T)

# Online calcresult = 74.36\% vs 74.3567\%
calculate_iga_risk_postbiopsy(egfr = 50, sbp=150, dbp=90,  proteinuria=1.2, age=30,
                              race="Japanese", RASB=T,
                              mest_M=1,mest_E=1,mest_S=1,mest_T=1, time=60)

# Online calcresult = 37.01\% vs 37.01\%
calculate_iga_risk_postbiopsy(egfr = 50, sbp=150, dbp=90,  proteinuria=1.2, age=30,
                              race="Caucasian", RASB=T,
                              mest_M=1,mest_E=1,mest_S=1,mest_T=1, time=60)

# Online calcresult = 49.69\% vs 27.98\% # ONLINE uses no race if "Other" is specified
calculate_iga_risk_postbiopsy(egfr = 50, sbp=150, dbp=90,  proteinuria=1.2, age=30,
                              race="Other", RASB=T,
                              mest_M=1,mest_E=1,mest_S=1,mest_T=1, time=60)

# Online calcresult = 49.69\% vs 49.69\%
calculate_iga_risk_postbiopsy_norace(egfr = 50, sbp=150, dbp=90,  proteinuria=1.2, age=30,
                                     RASB=T,
                                     mest_M=1,mest_E=1,mest_S=1,mest_T=1, time=60)
calculate_iga_risk_postbiopsy_norace(egfr = 50, sbp=150, dbp=90,  proteinuria=1.2, age=30,
                                     RASB=T,
                                     mest_M=1,mest_E=1,mest_S=1,mest_T=1, time=60,
                                     explanation = T)
}
