% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_isi_avignon.R
\name{calculate_isi_avignon}
\alias{calculate_isi_avignon}
\title{Avignon Insulin Sensitivity Index Calculation (from standard 75g OGTT)}
\usage{
calculate_isi_avignon(
  time,
  glucose,
  insulin,
  weight = NA,
  time_units = "min",
  glucose_units = "mg/dl",
  insulin_units = "uU/ml",
  weight_units = "kg"
)
}
\arguments{
\item{time}{Vector of time values (in minutes)}

\item{glucose}{Vector of glucose values (in mg/dL)}

\item{insulin}{Vector of insulin values (in uU/mL)}

\item{weight}{Weight (kg)}

\item{time_units}{if units are not in "min", can indicate here for unit
conversion (options "min" or "hr")}

\item{glucose_units}{if units are not in "mg/dl", can indicate here for unit
conversion (options "mg/dl" or "mmol/l")}

\item{insulin_units}{if units are not in "uU/ml", can indicate here for unit
conversion (options "uU/ml" or "pmol/l")}

\item{weight_units}{weight units, if not in kg}
}
\value{
A dataframe containing 3 values for basal, 2hr and mean Insulin
Sensitivity values
}
\description{
This function calculates the Avignon Insulin Sensitivity Index using glucose
and insulin sampled during a standard 75g oral glucose tolerance test.
Methods described in \href{https://pubmed.ncbi.nlm.nih.gov/10375055/}{Avignon
et al.}
}
\details{
Standard timepoints are 0, 30, 60, 90, and 120 min. Note: insulin unit
conversion may differ differ depending on assay. Insulin (pmol/l) = insulin
(uU/ml)*6

\code{calculate_isi_avignon()} accepts 3 separate vectors for time, glucose,
insulin and a single weight value.
}
\examples{
# individual objects for each item
time=c(0, 30, 60, 90, 120)              # minutes
glucose=c(93, 129, 178, 164, 97)        # mg/dL
insulin=c(12.8, 30.7, 68.5, 74.1, 44.0) # uU/mL
calculate_isi_avignon(time, glucose, insulin, weight=70)

# handling data stored in a dataframe
ogtt1 <- data.frame(time=c(0, 30, 60, 90, 120),              # minutes
                    glucose=c(93, 129, 178, 164, 97),        # mg/dL
                    insulin=c(12.8, 30.7, 68.5, 74.1, 44.0)) # uU/mL
calculate_isi_avignon(ogtt1$time, ogtt1$glucose, ogtt1$insulin) #

# example from Gutch et al 2015
ogtt2 <- data.frame(time=c(0, 30, 60, 90, 120),              # minutes
                    glucose=c(100, 160, 160, 160, 140),      # mg/dL
                    insulin=c(5, 10, 10, 10, 5))             # uU/mL
calculate_isi_avignon(ogtt2$time, ogtt2$glucose, ogtt2$insulin) #
calculate_isi_matsuda(ogtt2$time, ogtt2$glucose, ogtt2$insulin)

# Convert units
ogtt5 <- data.frame(time = c(0,0.5,1,1.5,2), # time in hours
                    glucose = c(5.167, 7.167, 9.889, 9.111, 5.3889), # glucose in mmol/l
                    insulin = c(76.8,184.2,411,444.6,264)) # insulin in pmol/l

calculate_isi_avignon(time =  ogtt5$time,
                        glucose = ogtt5$glucose,
                        insulin = ogtt5$insulin, weight=80,
                        time_units = "hr", insulin_units = "pmol/l", glucose_units = "mmol/l")
}
