% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_isi_belfiore.R
\name{calculate_isi_belfiore}
\alias{calculate_isi_belfiore}
\title{Belfiore Insulin Sensitivity Index Calculation (from standard 75g OGTT)}
\usage{
calculate_isi_belfiore(
  time,
  glucose,
  insulin,
  time_units = "min",
  glucose_units = "mg/dl",
  insulin_units = "uU/ml"
)
}
\arguments{
\item{time}{a column name (unquoted) indicating time values (in minutes)}

\item{glucose}{a column name (unquoted) storing glucose values (in mg/dL)}

\item{insulin}{a column name (unquoted) storing insulin values (in uU/mL)}

\item{time_units}{if units are not in "min", can indicate here for unit
conversion (options "min" or "hr")}

\item{glucose_units}{if units are not in "mg/dl", can indicate here for unit
conversion (options "mg/dl" or "mmol/l")}

\item{insulin_units}{if units are not in "uU/ml", can indicate here for unit
conversion (options "uU/ml" or "pmol/l")}
}
\value{
Insulin Sensitivity-Gutt (mL/kg/min) as a single value
}
\description{
This function calculates the Belfiore Insulin Sensitivity Index using glucose
and insulin sampled during a standard 75g oral glucose tolerance test.
Methods described in \href{https://pubmed.ncbi.nlm.nih.gov/9562967/}{Belfiore
et al.}.
}
\details{
Note the formula uses a population normalization factor representing a Normal
average value for Glucose x Insulin during the OGTT, which does not have a
commonly accepted standard. 

Standard timepoints are 0, 60, and 120 min. Note: insulin unit conversion may
differ differ depending on assay. Insulin (pmol/l) = insulin (uU/ml)*6

`calculate_isi_belfiore()` accepts 3 separate vectors for time, glucose,
insulin.
}
\examples{
# individual objects for each item
time=c(0, 30, 60, 90, 120)              # minutes
glucose=c(93, 129, 178, 164, 97)        # mg/dL
insulin=c(12.8, 30.7, 68.5, 74.1, 44.0) # uU/mL
calculate_isi_belfiore(time, glucose, insulin, bmi=30)

# handling data stored in a dataframe
ogtt1 <- data.frame(time=c(0, 30, 60, 90, 120),              # minutes
                    glucose=c(93, 129, 178, 164, 97),        # mg/dL
                    insulin=c(12.8, 30.7, 68.5, 74.1, 44.0)) # uU/mL
calculate_isi_belfiore(ogtt1$time, ogtt1$glucose, ogtt1$insulin, bmi=30)

# example from Gutch et al 2015
ogtt2 <- data.frame(time=c(0, 30, 60, 90, 120),              # minutes
                    glucose=c(100, 160, 160, 160, 140),      # mg/dL
                    insulin=c(5, 10, 10, 10, 5))             # uU/mL
calculate_isi_belfiore(ogtt2$time, ogtt2$glucose, ogtt2$insulin, bmi=30)
calculate_matsuda_index(ogtt2$time, ogtt2$glucose, ogtt2$insulin)

# Convert units
ogtt5 <- data.frame(time = c(0,0.5,1,1.5,2), # time in hours
                    glucose = c(5.167, 7.167, 9.889, 9.111, 5.3889), # glucose in mmol/l
                    insulin = c(76.8,184.2,411,444.6,264)) # insulin in pmol/l

calculate_isi_belfiore(time =  ogtt5$time,
                        glucose = ogtt5$glucose,
                        insulin = ogtt5$insulin, bmi=30,
                        time_units = "hr", insulin_units = "pmol/l", glucose_units = "mmol/l")
}
