% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lvmass.R
\name{calculate_lvmass}
\alias{calculate_lvmass}
\title{Calculate LV Mass using Echocardiogram results}
\usage{
calculate_lvmass(lvedd, ivsd, lvpwd, bsa, length_units = "cm")
}
\arguments{
\item{lvedd}{LV Internal Diameter Diastole}

\item{ivsd}{LV Septal Thickness Diastole}

\item{lvpwd}{LV Posterior Thickness Diastole}

\item{bsa}{Body surface area (\eqn{m^2})}

\item{length_units}{}
}
\value{
dataframe of echocardiagraphic LV mass results
}
\description{
Uses methods according to \href{https://pubmed.ncbi.nlm.nih.gov/25559473/}{American Society of Echocardiography}
}
\examples{
calculate_lvmass(lvedd = 50.5, ivsd = 10.7, lvpwd = 13.3, bsa=2.42, length_units = "mm")
}
