% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_matsuda_index_df.R
\name{calculate_matsuda_index_df}
\alias{calculate_matsuda_index_df}
\title{Matsuda-Defronzo Index Calculation (from standard 75g OGTT)}
\usage{
calculate_matsuda_index_df(
  df,
  timeCol = time,
  glucoseCol = glucose,
  insulinCol = insulin,
  time_units = "min",
  glucose_units = "mg/dl",
  insulin_units = "uU/ml"
)
}
\arguments{
\item{df}{A long-format data frame}

\item{timeCol}{a column name (unquoted) indicating time values (in minutes)}

\item{glucoseCol}{a column name (unquoted) storing glucose values (in mg/dL)}

\item{insulinCol}{a column name (unquoted) storing insulin values (in uU/mL)}

\item{time_units}{if units are not in "min", can indicate here for unit conversion (options "min" or "hr")}

\item{glucose_units}{if units are not in "mg/dl", can indicate here for unit conversion (options "mg/dl" or "mmol/l")}

\item{insulin_units}{if units are not in "uU/ml", can indicate here for unit conversion (options "uU/ml" or "pmol/l")}
}
\value{
Matsuda index as a single value
}
\description{
This function calculates the Matsuda index using glucose and insulin 
sampled during a standard 75g oral glucose tolerance test.
Results can be checked using the \href{http://mmatsuda.diabetes-smc.jp/MIndex.html}{Matsuda online calculator}.
}
\details{
Standard timepoints are 0, 30, 60, 90, and 120 min.
Note: insulin unit conversion may differ differ depending on assay. 
Insulin (pmol/l) = insulin (uU/ml)*6 

`calculate_matsuda_index_df()` requires a dataframe input with time, glucose, insulin columns.
}
