% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mesa_chdrisk.R
\name{calculate_mesa_chdrisk}
\alias{calculate_mesa_chdrisk}
\title{Calculate MESA CHD Risk (with CAC score)}
\usage{
calculate_mesa_chdrisk(
  age,
  gender,
  race,
  cac = NA,
  Tc,
  HDL,
  SBP,
  current_smoker = FALSE,
  using_antihypertensive_medication = FALSE,
  using_lipidloweringagent = FALSE,
  diabetes = FALSE,
  MI_fmhx = FALSE,
  chol_units = "mg/dL"
)
}
\arguments{
\item{age}{Age in years}

\item{gender}{Gender, (Female/Male)}

\item{race}{Race, one of (White, Black, Hispanic, Asian)}

\item{cac}{CAC score (Agatston score). If not available, set to NA.}

\item{Tc}{Total Cholesterol (mg/dL). converted to mmol/L internally}

\item{HDL}{HDL  Cholesterol (mg/dL). converted to mmol/L internally}

\item{SBP}{Systolic Blood Pressure, mmHg}

\item{current_smoker}{current_smoker, T/F}

\item{using_antihypertensive_medication}{HTN medication use, T/F}

\item{using_lipidloweringagent}{Lipid Lowering Medication use, T/F}

\item{diabetes}{Diabetes, T/F}

\item{MI_fmhx}{Family history of MI (in first degree relative), T/F}

\item{chol_units}{default = mg/dL; cholesterol units not in mg/dL then define
here for conversion}
}
\value{
10-year MESA CHD Risk (percent) in Dataframe format
}
\description{
Estimates 10-year risk of ASCVD events as described in the MESA study. Full
description provided in \href{http://www.ncbi.nlm.nih.gov/pubmed/26449133}{McClelland  et al.,
JACC 2015}. An online calculator can be found at the \href{https://internal.mesa-nhlbi.org/about/procedures/tools/mesa-score-risk-calculator}{MESA home website}. CHD Risk Scorewas  defined as a composite outcome including nonfatal myocardial infarctions, resuscitated cardiac arrest, probable angina, definite angina followed by revascularization, and fatal CHD.
}
\examples{
calculate_mesa_chdrisk(age=50, cac = 0,
                       gender = "male",
                       race = "white",
                       Tc=200, 
                       HDL=34, 
                       SBP=125,  
                       current_smoker=FALSE,
                       using_antihypertensive_medication=FALSE,
                       using_lipidloweringagent=FALSE,
                       diabetes=FALSE,
                       MI_fmhx=FALSE)

calculate_mesa_chdrisk(age=60, # case 
                       cac = 220,
                       gender = "male",
                       race = "white",
                       Tc=200, 
                       HDL=45, 
                       SBP=126,  
                       current_smoker=FALSE,
                       using_antihypertensive_medication=FALSE,
                       using_lipidloweringagent=FALSE,
                       diabetes=FALSE,
                       MI_fmhx=TRUE)
calculate_mesa_chdrisk(age=60, # average healthy 60yo
                       cac = NA,
                       gender = "male",
                       race = "white",
                       Tc=192, 
                       HDL=45, 
                       SBP=123,  
                       current_smoker=FALSE,
                       using_antihypertensive_medication=FALSE,
                       using_lipidloweringagent=FALSE,
                       diabetes=FALSE,
                       MI_fmhx=TRUE)
}
