% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mgfr_msp.R
\name{calculate_mgfr_msp}
\alias{calculate_mgfr_msp}
\title{Iohexol Multiple Sample Protocol (MSP) estimate of GFR}
\usage{
calculate_mgfr_msp(
  time,
  iohexol,
  omnipaque_v = 300,
  ioh_inj_wt = NULL,
  ioh_inj_vol = NULL,
  height,
  weight,
  height_units = "cm",
  weight_units = "kg"
)
}
\arguments{
\item{time}{A vector of time values (minutes)}

\item{iohexol}{A vector of Iohexol concentrations (mcg/mL) determined at the
provided times. These values should not be log transformed}

\item{omnipaque_v}{Omnipaque version (eg 300 = Omnipaque 300 recommended).
Other versions commonly available include 300 and 350 (USA) and 140, 280,
240 (other areas)}

\item{ioh_inj_wt}{Iohexol injection Weight by syringe weight determination
(Pre-Post weight difference; recommnded method) in grams}

\item{ioh_inj_vol}{Iohexol injection Volume by syringe volume injected (Not
preferred; use if weights not available)}

\item{height}{Patient Height, cm}

\item{weight}{Patient Weight, kg}

\item{height_units}{Height units, if not in cm}

\item{weight_units}{Weight units, if not in kg}
}
\value{
Data.frame containing results with calculated values for the
  following variables:

*  `iohexol_m` Iohexol dose (mcg Iodine)
*  `iohexol_0` Iohexol intercept (mcg/mL)
*  `iohexol_k` Iohexol elimination constant (k)
*  `rsq` Iohexol model R-sqaured value (recommend > 0.97)
*  `ioh_auc` Iohexol area under the curve (AUC, mcg min / mL)
*  `ioh_CL` Iohexol Clearance (dose/area)
*  `mGFR_1c` GFR (mL/min) determined by Iohexol clearance
*  `mGFR_1c_bsa` GFR indexed to BSA (mL/min/1.73 m2 bsa)
}
\description{
This method uses multiple sample measurements of iohexol to determine GFR
(glomerular filtration rate), using a single compartment model of the late
slope and adjustment using the Brochner-Mortenson correction equation.
Relevant Methods and recommendations are detailed in the
\href{https://pubmed.ncbi.nlm.nih.gov/39097002/}{European Kidney Function
Consortium statment paper}.
The function can be compared against an excel spreadsheet provided in the
supplemental material.
}
\examples{
XLSX example data
df1_dem <- data.frame(height=168,     # cm
                      weight=87,      # kg
                      ioh_inj_wt = 6.82594, # injected weight g; syringe wt pre-post
                      ioh_vol = 5.06, # mL
                      ioh_conc = 647)   # Omnipaque 300

dat <- data.frame(time=c(160,180,200,220,232,240),
                  iohexol_ucg_ml = c(70,60,47,37,30,25)) 

calculate_mgfr_msp(dat$time, dat$iohexol_ucg_ml,
                   ioh_inj_vol = 5.06,
                   height = df1_dem$height, weight = df1_dem$weight)  
calculate_mgfr_msp(dat$time, dat$iohexol_ucg_ml,
                   ioh_inj_wt = 6.82594,
                   height = df1_dem$height, weight = df1_dem$weight)  

# Example for multiple dataset analysis
df2_dem <- data.frame(id=1:5,
                      height= rnorm(5, 168, 10),     # cm
                     weight= rnorm(5, 87, 5),      # kg
                     ioh_inj_wt = rep(6.82594,5), # injected weight g; syringe wt pre-post
                     ioh_vol = rep(5.06,5)) # mL
dat2 <- data.frame(id=rep(1:5, each=6),
                  time = rep(c(160,180,200,220,232,240), 5),
                  iohexol_ucg_ml = rep(c(70,60,47,37,30,25), 5))

# merged with time series df in a column
df2_m <- merge(df2_dem, dat2, all.x = T) |> 
 dplyr::group_by(id) |> 
 tidyr::nest(dat = c(time, iohexol_ucg_ml))

df2_m |> 
 mutate(mgfr = map(dat, ~calculate_mgfr_msp(.x$time, .x$iohexol_ucg_ml,
                                            ioh_inj_vol = ioh_vol,
                                            height = height, weight = weight))) |> 
 tidyr::unnest(mgfr)
}
