% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mgfr_relm.R
\name{calculate_mgfr_relm}
\alias{calculate_mgfr_relm}
\title{Relmipirazin measured GFR with 2-compartment model (early and late sampling)}
\usage{
calculate_mgfr_relm(
  time,
  relmipirazin_conc,
  relm_units = "ng/mL",
  time_units = "min",
  id = NULL,
  height = NA,
  height_units = "m",
  weight = NA,
  weight_units = "kg",
  nls_weights = TRUE,
  output = "summary"
)
}
\arguments{
\item{time}{A vector of time values (minutes)}

\item{relmipirazin_conc}{A vector of relmipirazin plasma measurements (ug/mL)}

\item{relm_units}{relmipirazin concentration units, defaults to \code{ng/mL}}

\item{time_units}{Time units, defaults to \code{min}}

\item{id}{Study identifier (optional, passed to plot title)}

\item{height}{Patient Height, cm}

\item{height_units}{Height units, if not in cm}

\item{weight}{Patient Weight, kg}

\item{weight_units}{Weight units, if not in kg}

\item{output}{Desired output, defaults to \code{summary} of model. Alternatively
can specify \code{gfr}, \code{gfr_bsa}, \code{fit}, or \code{plot}

\code{summary} results include the named variables:
\itemize{
\item \code{mgfr_2c} The measured GFR (mL/min)
\item \code{mgfr_2c_bsa} The measured GFR (mL/min/1.73 m2) indexed to body surface area (BSA calculated by DuBois equation)
\item \code{relmipirazin_m}   Relmipirazin mass (ucg) injected. Determined either by user-provided syringe weight or volume injected.
\item \code{relmipirazin_0}   Relmipirazin concentration at t=0 calculated by A+B (theoretical, not measured)
\item \code{relmipirazin_vd}  Relmipirazin Volume of distribution estimated by Dose/relmipirazin\link{t0}
\item \code{relm_auc}  Relmipirazin calculated AUC from t=0 to Infinity, estimated by A/a + B/b
\item \code{A} Model parameter A
\item \code{a} Model parameter a, or \eqn{\alpha}
\item \code{B} Model parameter B
\item \code{b} Model parameter b, or \eqn{\beta}
\item \code{model_r2} Model Pseudo-R2, calculated as the linear regression R2 for predicted~observed values
\item \code{k10} Relmipirazin elimination rate constant from central compartment (1/min)
\item \code{k21} Relmipirazin transfer rate constant from peripheral to central compartment (1/min)
\item \code{k12} Relmipirazin transfer rate constant from central to peripheral compartment (1/min)
}

\code{gfr} returns a single \code{mgfr_2c} value, the measured GFR (mL/min).
\code{gfr_bsa} returns a single \code{mgfr_2c_bsa} value, the measured GFR
(mL/min/1.73 m2) indexed to body surface area (BSA calculated by DuBois
equation).

\code{fit} returns the non-linear regression model fit object.

\code{plot} returns a base-R plot of observed and predicted regression curve vs
time, and summary measures in the legend.}

\item{relm_inj_wt}{Relmipirazin injection Weight by syringe weight determination
(Pre-Post weight difference; recommended method) in grams}

\item{relm_inj_vol}{Relmipirazin injection Volume by syringe volume injected (Not
preferred; use if weights not available)}
}
\value{
Desired output with either a data.frame of results (\code{summary}), a
single value of BSA adjusted GFR (\code{gfr}), an \verb{nls object} for model fit
(\code{fit}), or a plot with observed values, model fit curve, and summary
results in the figure legend (\code{plot})
}
\description{
This method requires multiple early (alpha phase, typically <120 min) and
late (beta phase) sample measurements of relmipirazin to determine GFR (glomerular
filtration rate). This function uses nonlinear modeling via the \code{nls()}
function to fit the 2-compartment model of relmipirazin kinetics using the formula \eqn{C=A*e^(- \alpha * t) +
B*e^(- \beta * t) }. The general method and provided examples are
modified from \href{https://pubmed.ncbi.nlm.nih.gov/16612328/}{Schwartz et al.
Kidney Int. 2006.}. Results are returned as a non-indexed value for GFR (\code{mgfr_2c}) and also indexed to BSA of 1.73m2 if height and weight are provided (\code{mgfr_2c_bsa}).
The summary provides estimates of kinetic parameters (k10, k21, k12) which can be used in ODE models, but is not implemented here.
}
\examples{
calculate_mgfr_relm(df_rel_a$time, df_rel_a$relmapirazin_ng_ml, height = 1.6002, weight = 71.698, output="plot")
calculate_mgfr_relm(df_rel_a$time, df_rel_a$relmapirazin_ng_ml, height = 1.6002, weight = 71.698, output="plot", nls_weights = FALSE)

# ODE MODEL FOR COMPARISON; 
# Example only to verify ODE parameter estimates, not required for GFR calculation
# This model uses solution from the nls model to get the kinetic parameters
library(deSolve)
m2c <- function(Time, State, Pars) {
  with(as.list(c(State, Pars)), {
    drelmipirazin <- -k10*relmipirazin - k12*relmipirazin + k21*C2 # Central Conc
    dC2 <-  k12*relmipirazin - k21*C2                    # Peripheral Conc
    return(list(c(drelmipirazin, dC2)))
  })
}
# get parameters from nls model
relm_nlfit <- calculate_mgfr_2c(dat$time, dat$relmipirazin_ug_ml, height = 1.67, weight = 70, relm_inj_vol = 5)
pars  <- c(k10 = relm_nlfit$k10, k21 = relm_nlfit$k21, k12 = relm_nlfit$k12)
# plug into the ODE model solution
yini  <- c(relmipirazin = relm_nlfit$A+relm_nlfit$B, C2=0)
times <- seq(0, 360, by = 1)
out   <- ode(yini, times, m2c, pars)
out_df <- as.data.frame(out)
# Plot to verfiy kinetic parameters vs nls model
calculate_mgfr_2c(dat$time, dat$relmipirazin_ug_ml, height = 1.67, weight = 70, relm_inj_vol = 5, output="plot")
lines(out_df$time, out_df$relmipirazin, lty=2)
}
