% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mgfr_ss.R
\name{calculate_mgfr_ss}
\alias{calculate_mgfr_ss}
\title{Iohexol Single Sample (SS) measure of GFR}
\usage{
calculate_mgfr_ss(
  egfr,
  height,
  weight,
  sex,
  age,
  iohexol_m,
  iohexol_ss,
  time_ss,
  weight_units = "kg",
  height_units = "cm",
  ecv_method = "Faucon"
)
}
\arguments{
\item{egfr}{Model-based estimate of GFR (eg CKD-EPI eGFR). Used as a starting point for estimation.}

\item{height}{Height, cm}

\item{weight}{Weight, kg}

\item{sex}{Sex (M/F)}

\item{age}{Age, years}

\item{iohexol_m}{Iohexol dose injected (mcg)}

\item{iohexol_ss}{Iohexol concentration (mcg/mL)}

\item{time_ss}{Time of iohexol sample (minutes)}

\item{weight_units}{Weight units, if not in kg}

\item{height_units}{Height units, if not in cm}

\item{ecv_method}{ECV Formula (default =  "Faucon"; also can use "Granerus")}
}
\value{
data.frame with `mgfr_ss` (non-indexed) and `mgfr_ss_bsa` (GFR indexed to standard BSA of 1.73m2)
}
\description{
This method uses a single sample measurement of iohexol to determine GFR
(glomerular filtration rate), by optimizing 3 equations initially described
by \href{https://pubmed.ncbi.nlm.nih.gov/6684515/}{Jacobsson}. Relevant
Methods and recommendations are detailed in the
\href{https://pubmed.ncbi.nlm.nih.gov/39097002/}{European Kidney Function
Consortium statment paper}. The R script is modified from the supplemental methods supplied in that paper, which also supplies an excel spreadsheet which can be used for comparison. One minor difference is the use of a newer and more accurate formula to estimate ECV.
}
\examples{
calculate_mgfr_ss(egfr=40, 
                  height = 168, weight = 87, 
                  sex="mALE", age=80, 
                  iohexol_m = 3273820,
                  iohexol_ss= 60,
                  time_ss=390)
                  
# used alternate ECV calculation (doesn't make much difference)
calculate_mgfr_ss(egfr=40, 
                  height = 168, weight = 87, 
                  sex="male", age=80, ecv_method = "Granerus",
                  iohexol_m = 3273820,
                  iohexol_ss= 60,
                  time_ss=390)

}
\seealso{
[calculate_ecv()] for ECV calculation
 [calculate_bsa()] for BSA calculation
}
