% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_prevent_risk.R
\name{calculate_prevent_risk}
\alias{calculate_prevent_risk}
\title{Calculate AHA PREVENT Risk}
\usage{
calculate_prevent_risk(
  risk,
  gender,
  age,
  Tc,
  HDL,
  SBP,
  eGFR,
  BMI = NA,
  current_smoker = FALSE,
  using_antihypertensive_medication = FALSE,
  using_statin = FALSE,
  diabetes = FALSE,
  chol_units = "mg/dL"
)
}
\arguments{
\item{risk}{Desired Risk Calculation ("cvd", "ascvd", "hf", "cad", "stroke")}

\item{gender}{Gender, (Female/Male)}

\item{age}{Age in years}

\item{Tc}{Total Cholesterol (mg/dL). converted to mmol/L internally}

\item{HDL}{HDL  Cholesterol (mg/dL). converted to mmol/L internally}

\item{SBP}{Systolic Blood Pressure, mmHg}

\item{eGFR}{estimated GFR (ml/min/1.73m2)}

\item{BMI}{Body Mass Index (kg/m2), used only in HF estimate}

\item{current_smoker}{current_smoker, T/F}

\item{using_antihypertensive_medication}{HTN medication use, T/F}

\item{using_statin}{Statin use, T/F}

\item{diabetes}{Diabetes, T/F}

\item{chol_units}{default = mg/dL; cholesterol units not in mg/dL then define
here for conversion}
}
\value{
10- and 30-year Risk (percent) in Dataframe format
}
\description{
Uses the PREVENT Base model (sex-specific, race-free) risk equations: PREVENT
(AHA Predicting Risk of CVD Events). The PREVENT equations enable 10 and
30-year risk estimates for total CVD (composite of atherosclerotic CVD and
heart failure), ASCVD (atherosclerotic CVD), Heart failure (HF), Coronary
Artery Disease (CAD), and Stroke. Details
provided in \href{https://pubmed.ncbi.nlm.nih.gov/37947085/}{Khan et al.,
Circulation. 2024}. Extended models incorporating urine albumin, A1C and socioeconomic risk (by Zip code) are available.
}
\details{
Allowable range for Risk variables are:
\itemize{
\item Age 30-79
\item Total Cholesterol 130-320 mg/dL
\item HDL Cholesterol 20-100 mg/dL
\item SBP 90-200 mmHg
\item BMI 18.5 - 40 mg/m^2
\item eGFR 15-150 mL/min/1.73m^2
\item Urine albumin/Creatinine 0.1-2500 mg/g Creatinine
}
}
\examples{

# Example results below have been validated against original publication (XLSX file with calc check)
# women
calculate_prevent_risk(risk="cvd",gender="Female", age=50, Tc=200, HDL=45, SBP=160, eGFR=90,
                      using_antihypertensive_medication = TRUE, diabetes = TRUE)
calculate_prevent_risk(risk="ascvd",gender="female", age=50, Tc=200, HDL=45, SBP=160, eGFR=90,
                      using_antihypertensive_medication = TRUE, diabetes = TRUE)
calculate_prevent_risk(risk="hf",gender="female", age=50, Tc=200, HDL=45, SBP=160, eGFR=90, BMI=35,
                      using_antihypertensive_medication = TRUE, diabetes = TRUE)
calculate_prevent_risk(risk="cad",gender="female", age=50, Tc=200, HDL=45, SBP=160, eGFR=90,
                      using_antihypertensive_medication = TRUE, diabetes = TRUE)
calculate_prevent_risk(risk="stroke",gender="female", age=50, Tc=200, HDL=45, SBP=160, eGFR=90,
                      using_antihypertensive_medication = TRUE, diabetes = TRUE)

# men
calculate_prevent_risk(risk="cvd",gender="male", age=50, Tc=200, HDL=45, SBP=160, eGFR=90,
                      using_antihypertensive_medication = TRUE, diabetes = TRUE)
calculate_prevent_risk(risk="ascvd",gender="male", age=50, Tc=200, HDL=45, SBP=160, eGFR=90,
                      using_antihypertensive_medication = TRUE, diabetes = TRUE)
calculate_prevent_risk(risk="hf",gender="male", age=50, Tc=200, HDL=45, SBP=160, eGFR=90, BMI=35,
                      using_antihypertensive_medication = TRUE, diabetes = TRUE)
calculate_prevent_risk(risk="cad",gender="male", age=50, Tc=200, HDL=45, SBP=160, eGFR=90,
                      using_antihypertensive_medication = TRUE, diabetes = TRUE)
calculate_prevent_risk(risk="stroke",gender="male", age=50, Tc=200, HDL=45, SBP=160, eGFR=90,
                      using_antihypertensive_medication = TRUE, diabetes = TRUE)
}
