% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_quicki.R
\name{calculate_quicki}
\alias{calculate_quicki}
\title{Quick Index of Insulin Sensitivity (QUICKI)}
\usage{
calculate_quicki(
  glucose,
  insulin,
  glucose_units = "mg/dl",
  insulin_units = "uU/ml"
)
}
\arguments{
\item{glucose}{fasting glucose (mg/dL)}

\item{insulin}{fasting insulin (uU/mL)}

\item{glucose_units}{if units are not in "mg/dl", can indicate here for unit conversion (options "mg/dl" or "mmol/l")}

\item{insulin_units}{if units are not in "uU/ml", can indicate here for unit conversion (options "uU/ml" or "pmol/l")}
}
\value{
QUICKI estimate of insulin sensitivity, as a single numeric value
}
\description{
Quick Index of Insulin Sensitivity (QUICKI)
}
\examples{
calculate_quicki(glucose = c(93), insulin = c(12.8))
calculate_quicki(93, 12.8)
calculate_quicki(93, NA) # error
calculate_quicki(5.167, 76.8, glucose_units = "mmol/l", insulin_units = "pmol/l")
}
