% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_stumvoll_beta.R
\name{calculate_stumvoll_beta}
\alias{calculate_stumvoll_beta}
\title{Stumvoll beta cell function Calculation (from standard 75g OGTT)}
\usage{
calculate_stumvoll_beta(
  time,
  glucose,
  insulin,
  time_units = "min",
  glucose_units = "mg/dl",
  insulin_units = "uU/ml"
)
}
\arguments{
\item{time}{a column name (unquoted) indicating time values (in minutes)}

\item{glucose}{a column name (unquoted) storing glucose values (in mg/dL)}

\item{insulin}{a column name (unquoted) storing insulin values (in uU/mL)}

\item{time_units}{if units are not in "min", can indicate here for unit
conversion (options "min" or "hr")}

\item{glucose_units}{if units are not in "mg/dl", can indicate here for unit
conversion (options "mg/dl" or "mmol/l")}

\item{insulin_units}{if units are not in "uU/ml", can indicate here for unit
conversion (options "uU/ml" or "pmol/l")}
}
\value{
Stumvoll beta cell estimates (\code{stumvoll_ph1}, \code{stumvol_ph2}) values as a \code{data.frame}
}
\description{
This function calculates the Stumvoll phase 1 (ph1) and phase 2 (ph2) using glucose
and insulin sampled during the first 30 minutes of a standard 75g oral glucose tolerance test. Described in \href{https://pubmed.ncbi.nlm.nih.gov/10868854/}{Stumvoll et al.}
}
\details{
Standard timepoints are 0, 30, 60, 90, and 120 min. Note: insulin unit
conversion may differ differ depending on assay. Insulin (pmol/l) = insulin
(uU/ml)*6

\code{calculate_stumvoll_beta()} accepts 3 separate vectors for time, glucose,
insulin.
}
\examples{
# individual objects for each item
time=c(0, 30, 60, 90, 120)              # minutes
glucose=c(93, 129, 178, 164, 97)        # mg/dL
insulin=c(12.8, 30.7, 68.5, 74.1, 44.0) # uU/mL
calculate_stumvoll_beta(time, glucose, insulin)
}
