% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_uacr_from_ua.R
\name{calculate_uacr_from_ua}
\alias{calculate_uacr_from_ua}
\title{Calculate Urine Albumin/Creatinine ratio from Dipstick Urinalysis Protein}
\usage{
calculate_uacr_from_ua(ua_protein, gender, diabetic = NA, hypertensive = NA)
}
\arguments{
\item{ua_protein}{Dipstick UA protein result (negative, trace, +, ++, +++)}

\item{gender}{Gender (Female/Male)}

\item{diabetic}{Diabetic (T/F)}

\item{hypertensive}{Hypertensive (T/F) using prior cutoff of >140/90 or on
meds}

\item{PCR_units}{urine Protein/creatinine units ("mg/g" default)}
}
\value{
Dataframe with predicted ACR (pACR) and estimated error
}
\description{
Estimated urine albumin creatinine ratio by converting urine
protein/creatinine measurement, or by estimation from dipstick urinalysis
(negative, trace, +, ++, +++). methods described in
\href{https://pubmed.ncbi.nlm.nih.gov/32658569/}{Sumida et al. Annals Int
Med. 202}.
}
\examples{
# calculate_uacr_from_ua(ua_protein = "trace", gender = "female", diabetic = T, hypertensive = T)
calculate_uacr_from_ua(ua_protein = "+", gender = "female", diabetic = T, hypertensive = T)
calculate_uacr_from_ua(ua_protein = "++", gender = "female", diabetic = T, hypertensive = T)
calculate_uacr_from_ua(ua_protein = "+++", gender = "female", diabetic = T, hypertensive = T)

calculate_uacr_from_ua(ua_protein = "trace", gender = "male", diabetic = T, hypertensive = T)
calculate_uacr_from_ua(ua_protein = "+", gender = "male", diabetic = T, hypertensive = T)
calculate_uacr_from_ua(ua_protein = "++", gender = "male", diabetic = T, hypertensive = T)
calculate_uacr_from_ua(ua_protein = "+++", gender = "male", diabetic = T, hypertensive = T)
}
