% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_uacr_from_upcr.R
\name{calculate_uacr_from_upcr}
\alias{calculate_uacr_from_upcr}
\title{Calculate Urine Albumin/Creatinine ratio from Protein/Creatinine ratio}
\usage{
calculate_uacr_from_upcr(
  PCR,
  gender,
  diabetic = NA,
  hypertensive = NA,
  PCR_units = "mg/g"
)
}
\arguments{
\item{PCR}{urine Protein/creatinine ratio (mg/g creat.)}

\item{gender}{Gender (Female/Male)}

\item{diabetic}{Diabetic (T/F)}

\item{hypertensive}{Hypertensive (T/F) using prior cutoff of >140/90 or on
meds}

\item{PCR_units}{urine Protein/creatinine units ("mg/g" default)}
}
\value{
Dataframe with predicted ACR (pACR) and estimated error
}
\description{
Estimated urine albumin creatinine ratio by converting urine
protein/creatinine measurement, or by estimation from dipstick urinalysis
(negative, trace, +, ++, +++). methods described in
\href{https://pubmed.ncbi.nlm.nih.gov/32658569/}{Sumida et al. Annals Int
Med. 202}.
}
\examples{
calculate_uacr_from_upcr(PCR = 1000, gender = "female", diabetic = T, hypertensive = T)
calculate_uacr_from_upcr(PCR = 500.0, gender = "female", diabetic = T, hypertensive = T)
calculate_uacr_from_upcr(PCR = 150, gender = "female", diabetic = T, hypertensive = T)
}
