% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_author_list.R
\name{clean_author_list}
\alias{clean_author_list}
\title{Reformat list of authors into Bibliography format}
\usage{
clean_author_list(string_authors)
}
\arguments{
\item{string_authors}{A text list of authors}
}
\value{
String of reformatted names
}
\description{
Reformats from typical list of full names (First middle Last,...) to standard
bibliographic format (Last FM, ...). Strips out numbers. Will not work for
some double and/or hyphenated names or other edge cases which will still have
to be cleaned manually.
}
\examples{
authors <- "Desmaré van Rooyen1,2, Sascha Bandulik3, Chandan Kumar-Sinha4,5, 
  Aaron M. Udager4,5,6, Chaelin Lee2, Heather Wachtel7, Debbie L. Cohen8, 
  James M. Luther9, Thomas Giordano6, Adina Turcu2, Richard Warth3, 
  William E. Rainey1,2 and Juilee Rege*1"
clean_author_list(authors)
auth <- "B. Ilkin Safa1, Jared Oakes1, Leslie Kirk1, Eric Olson1, 
Christian M. Warren1, Erin M. Wilfong1, Matthew T. Stier1, Samuel S. Bailin1,
 Curtis Gabriel1, Jeffrey C. Rathmell1, C. Robb Flynn1, J. Matthew Luther1, 
 John R. Koethe1,2, Spyros A. Kalams1, Celestine N. Wanjalla1, Mona Mashayekhi1"
clean_author_list(auth) 
}
