% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cholesterol_to_mM.R
\name{convert_cholesterol_to_mM}
\alias{convert_cholesterol_to_mM}
\title{Convert Cholesterol (Total, HDL, or LDL) from mg/dL to mmol/L}
\usage{
convert_cholesterol_to_mM(cholesterol, cholesterol_units = "mg/dL")
}
\arguments{
\item{cholesterol}{numeric vector with cholesterol values}

\item{cholesterol_units}{currently only mg/dl}
}
\value{
numeric vetor of cholesterol in mmol/l
}
\description{
LDL, HDL and Total cholesterol are usually reported in mg/dL units. Because
the value is the cholesterol content, all use the same unit conversion.
}
\examples{
convert_cholesterol_to_mM(200, "mg/dl")
convert_cholesterol_to_mM(100, "mg/dL")
convert_cholesterol_to_mM(100, "mg_dl")
convert_cholesterol_to_mM(100, "mg") # improper units
convert_cholesterol_to_mM(100, "mmol/l")
convert_cholesterol_to_mgdl(5.171968, "mmol/l")
}
