% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cpeptide_to_nM.R
\name{convert_cpeptide_to_}
\alias{convert_cpeptide_to_}
\alias{convert_cpeptide_to_nM}
\alias{convert_cpeptide_to_pM}
\alias{convert_cpeptide_to_ng_ml}
\title{Convert C-peptide concentration to uU/mL}
\usage{
convert_cpeptide_to_nM(cpeptide, cpeptide_units)

convert_cpeptide_to_pM(cpeptide, cpeptide_units)

convert_cpeptide_to_ng_ml(cpeptide, cpeptide_units)
}
\arguments{
\item{cpeptide}{numeric vector of cpeptide values in pmol/l}

\item{cpeptide_units}{character string of units (pmol/l or similar)}
}
\value{
numeric vector of cpeptide values in uU/mL
}
\description{
Convert C-peptide concentration to uU/mL
}
\examples{
convert_cpeptide_to_nM(1.0, cpeptide_units = "ng/ml")
convert_cpeptide_to_pM(1.0, cpeptide_units = "ng/ml")
convert_cpeptide_to_pM(1.0, cpeptide_units = "ng/ml")
convert_cpeptide_to_ng_ml(330, cpeptide_units = "pM")
convert_cpeptide_to_ng_ml(0.33, cpeptide_units = "nM")
convert_cpeptide_to_ng_ml(1, cpeptide_units = "ng/ml")
}
