% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_creatinine_to_uM.R
\name{convert_creatinine_to_uM}
\alias{convert_creatinine_to_uM}
\title{Convert Cretainine from mg/dL to umol/L}
\usage{
convert_creatinine_to_uM(creatinine, creatinine_units = "mg/dL")
}
\arguments{
\item{creatinine}{numeric vector with creatinine values}

\item{creatinine_units}{currently only mg/dl}
}
\value{
numeric vetor of creatinine in umol/l
}
\description{
Convert Cretainine from mg/dL to umol/L
}
\examples{
convert_creatinine_to_uM(1, "mg/dl")
convert_creatinine_to_uM(1, "mg/dL")
convert_creatinine_to_uM(1, "mg_dl")
convert_creatinine_to_uM(1, "mg") # improper units
convert_creatinine_to_uM(1, "umol/l") 

convert_creatinine_to_mgdl(88.4, "umol/l") 

}
