% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_glucose_to_mM.R
\name{convert_glucose_to_mM}
\alias{convert_glucose_to_mM}
\title{Convert glucose from mg/dLl to mmol/L}
\usage{
convert_glucose_to_mM(glucose, glucose_units)
}
\arguments{
\item{glucose}{numeric vector with glucose values}

\item{glucose_units}{currently only mg/dl}
}
\value{
numeric vetor of glucose in mmol/l
}
\description{
Convert glucose from mg/dLl to mmol/L
}
\examples{
convert_glucose_to_mM(100, "mg/dl")
convert_glucose_to_mM(100, "mg/dL")
convert_glucose_to_mM(100, "mg_dl")
convert_glucose_to_mM(100, "mg") # improper units
convert_glucose_to_mM(100, "mmol/l") 

convert_glucose_to_mgdl(8, "mmol/l") 

}
