% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_height_m.R
\name{convert_height_m}
\alias{convert_height_m}
\title{Convert height to m}
\usage{
convert_height_m(height, height_units)
}
\arguments{
\item{height}{height in cm, m, or inches}

\item{height_units}{units in "cm", "m", or "inches"}
}
\value{
vector of height in meters (m)
}
\description{
Convert height to m
}
\examples{
convert_height_m(70, "inches")
convert_height_m(c(70, 75, 65), "inches")
}
