% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_insulin_to_uU_ml.R
\name{convert_insulin_to_uU_ml}
\alias{convert_insulin_to_uU_ml}
\title{Convert Insulin concentration to uU/mL}
\usage{
convert_insulin_to_uU_ml(insulin, insulin_units)
}
\arguments{
\item{insulin}{numeric vector of insulin values in pmol/l}

\item{insulin_units}{character string of units (pmol/l or similar)}
}
\value{
numeric vector of insulin values in uU/mL
}
\description{
Convert Insulin concentration to uU/mL
}
\examples{
convert_insulin_to_uU_ml(50, "pmol/L")
convert_insulin_to_pM(8.3333, "uu/ml")
}
