% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_height_to_m.R,
%   R/convert_length_to_in.R, R/convert_length_to_m.R, R/convert_length_to_mm.R
\name{convert_height_to_m}
\alias{convert_height_to_m}
\alias{convert_length_to_in}
\alias{convert_length_to_m}
\alias{convert_length_to_mm}
\title{Convert length units}
\usage{
convert_height_to_m(height, height_units)

convert_length_to_in(length, length_units)

convert_length_to_m(length, length_units)

convert_length_to_mm(length, length_units)
}
\arguments{
\item{length}{length in cm, m, or inches}

\item{length_units}{input units in "cm", "m", or "inches"}
}
\value{
vector of length in desired units specified by function name
}
\description{
Convert length units
}
\examples{
convert_length_to_mm(1.03, "cm")
convert_length_to_mm(1, "in")
convert_length_to_mm(1, "m")
convert_length_to_mm(c(1.03, 1, 1), "cm")

convert_length_to_m(70, "inches")
convert_length_to_m(c(70, 75, 65), "inches")
}
