% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_length_to_mm.R
\name{convert_length_to_mm}
\alias{convert_length_to_mm}
\title{Convert length to mm}
\usage{
convert_length_to_mm(length, length_units)
}
\arguments{
\item{length}{length in cm, m, or inches}

\item{length_units}{units in "cm", "m", or "inches"}
}
\value{
vector of length in millimeters (mm)
}
\description{
Convert length to mm
}
\examples{
convert_length_to_mm(1.03, "cm")
convert_length_to_mm(1, "in")
convert_length_to_mm(1, "m")
convert_length_to_mm(c(1.03, 1, 1), "cm")
}
