% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_time_to_min.R
\name{convert_time_to_min}
\alias{convert_time_to_min}
\title{Convert time to minutes}
\usage{
convert_time_to_min(time, time_units)
}
\arguments{
\item{time}{numeric vector}

\item{time_units}{specify the units of current time values}
}
\value{
numeric vector
}
\description{
Convert time to minutes
}
\examples{
convert_time_to_min(1.5, "hr")
convert_time_to_min(1.5, "hrs")
convert_time_to_min(1.5, "hours")
convert_time_to_min(60, "secs")
convert_time_to_min(60, "s")
}
