% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_triglycerides_to_mM.R
\name{convert_triglycerides_to_mM}
\alias{convert_triglycerides_to_mM}
\title{Convert Triglycerides from mg/dL to mmol/L}
\usage{
convert_triglycerides_to_mM(triglycerides, triglycerides_units = "mg/dL")
}
\arguments{
\item{triglycerides}{numeric vector with triglycerides values}

\item{triglycerides_units}{currently only mg/dl}
}
\value{
numeric vetor of triglycerides in mmol/l
}
\description{
LDL, HDL and Total triglycerides are usually reported in mg/dL units.
}
\examples{
convert_triglycerides_to_mM(200, "mg/dl")
convert_triglycerides_to_mM(150, "mg/dL")
convert_triglycerides_to_mM(100, "mg_dl")
convert_triglycerides_to_mM(100, "mg") # improper units
convert_triglycerides_to_mM(100, "mmol/l")
convert_triglycerides_to_mgdl(1.693576, "mmol/l")
}
