% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_weight_kg.R
\name{convert_weight_kg}
\alias{convert_weight_kg}
\title{Convert weight to kg}
\usage{
convert_weight_kg(weight, weight_units)
}
\arguments{
\item{weight}{weight in lbs g, or kg}

\item{weight_units}{units as "lbs" "g", or "kg"}
}
\value{
vector of weight in kg units
}
\description{
Convert weight to kg
}
\examples{
convert_weight_kg(165, "lbs")
convert_weight_kg(c(165, NA), "lbs")
convert_weight_kg(70000, "g")
}
