% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpeptide_params.R
\name{cpeptide_params}
\alias{cpeptide_params}
\title{C-peptide parameters estimates from population data}
\usage{
cpeptide_params(
  age,
  gender = NA,
  height,
  weight,
  category = "normal",
  weight_units = "kg",
  height_units = "m"
)
}
\arguments{
\item{age}{Age in years}

\item{gender}{Gender}

\item{height}{Height in meters}

\item{weight}{Weight in kg}

\item{category}{Category ("Normal", "Obese", "NIDDM")}

\item{weight_units}{weight units, if not in kg}

\item{height_units}{height units, if not in meters}
}
\value{
Data frame with variables `cp_vd`, `cp_halflife_short`, `cp_halflife_long`,
  `cp_fraction`, `cp_bsa`, `cp_k12`, `cp_k21`, `cp_k10`
}
\description{
This function estimates C-peptide kinetic parameters from population data.
Based on the regression models from
\href{https://pubmed.ncbi.nlm.nih.gov/1551497/}{Van Cauter et al}.
Participants were considered "Obese" in this study if body weight was >15%
above ideal body weight, rather than a commonly used BMI cutoff.
}
\details{
The parameters can be used in calculations to estimate insulin secretion
rates, shown in the example.
}
\examples{
cpeptide_params(age=28.1, height = 1.744666, weight = 69.4, gender = "F", category = "normal")
cpeptide_params(age=28.1, height = 1.744666, weight = 69.4, gender = "M", category = "normal")
cpeptide_params(age=35.2, height = 1.678461, weight = 107.9, gender = "F", category = "obese")
cpeptide_params(age=35.2, height = 1.678461, weight = 107.9, gender = "M", category = "obese")

library(dplyr)
dat <- data.frame(id =1:10,
  age = rnorm(10, 50, sd=8),
  gender = sample(c("M", "F"), 10, replace = T),
  height = rnorm(10, 1.7, sd=0.2),
  weight = rnorm(10, 100, sd=20),
  cpeptide_0 =  abs(rnorm(10, 2.0, 0.91))  ) # ng/dL

dat |>
  mutate(ibw = calculate_ibw(height, gender, height_units = "m", weight_units = "kg"),
         ibw_pct = (weight/ibw-1)*100,
         ibw_15 = case_when(ibw_pct>15 ~ "obese", # if >15\% above IBW then Obese
                            TRUE ~ "normal"),
         cpeptide_0_pmol = convert_cpeptide_to_pM(cpeptide_0, cpeptide_units="ng/ml"),
         cpeptide_params(age, gender, height, weight, category = ibw_15),
         # Basal Insulin Secretion Rate (pmol/min)
         ISR_0 = cp_k10*cpeptide_0_pmol*cp_vd)
}
