% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epo_nomogram.R
\name{epo_nomogram}
\alias{epo_nomogram}
\title{Erythropoietin (EPO) Nomogram Plot
Plots a single matched hematocrit, EPO data pair over the historical EPO nomogram. Data was taken from \href{https://pubmed.ncbi.nlm.nih.gov/2017231/}{Erslev. NEJM. 1991}}
\usage{
epo_nomogram(hematocrit, epo)
}
\arguments{
\item{hematocrit}{Hematocrit (\%) that corresponds to EPO}

\item{epo}{Serum Erythropoietin (mU/mL)}
}
\value{
a ggplot of datapoint over the EPO nomogram plot
}
\description{
Erythropoietin (EPO) Nomogram Plot
Plots a single matched hematocrit, EPO data pair over the historical EPO nomogram. Data was taken from \href{https://pubmed.ncbi.nlm.nih.gov/2017231/}{Erslev. NEJM. 1991}
}
\examples{
epo_nomogram(35, 12)
}
