% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_bp.R
\name{extract_bp}
\alias{extract_bp}
\title{Extract blood pressure values from text (SBP/DBP)}
\usage{
extract_bp(text, meanBP = TRUE, digits = 3)
}
\arguments{
\item{text}{A string containing blood pressure readings}

\item{meanBP}{if FALSE, does not return calculated MAP}

\item{digits}{Desired number of signficant digits, default=3}
}
\value{
A dataframe containing SBP, DBP and MAP
}
\description{
Separates SBP and DBP from a string of blood pressure readings, and returns
it as a dataframe. Mean arterial pressure is added unless default `meanBP`
option changed.
}
\examples{
bpreadings <- c("120/80", "90/50", "220/40")
bpreadings2 <- c("120 / 80", "90 / 50", "220 / 40")
extract_bp(bpreadings)
extract_bp(bpreadings2)
extract_bp(bpreadings2, meanBP = F)
}
