% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_library.R
\name{extract_library}
\alias{extract_library}
\title{Extract library annotation}
\usage{
extract_library(df, reverse_complement = FALSE)
}
\description{
This function extracts the relevant information from a screening library package for crispr-nf
}
