% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_numeric_part.R
\name{extract_numeric_part}
\alias{extract_numeric_part}
\title{Extract the numerical part from the count matrix}
\usage{
extract_numeric_part(counts)
}
\description{
Helper function to extract the numerical part of the matrix
}
