% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_glucose_fasting.R
\name{factor_glucose_fasting}
\alias{factor_glucose_fasting}
\title{Fasting Glucose Classification}
\usage{
factor_glucose_fasting(x)
}
\arguments{
\item{x}{A glucose value, or vector of values, in mg/dL. Converts
a number entered as a string to numeric if possible, otherwise returns \code{NA}}
}
\value{
A Factor for OGTT classification of fasting and 2hr glucose
}
\description{
\itemize{
\item \code{factor_glucose_fasting()} classifies fasting glucose as Normal (NFG),
Impaired (IFG), or Diabetes (DM2)
}
}
\examples{
factor_glucose_fasting(155)
factor_glucose_fasting(c(120, 130, 150, 200))
factor_glucose_fasting(c(120, 130, NA, 200))
}
