% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_glucose_tolerance.R
\name{factor_glucose_tolerance}
\alias{factor_glucose_tolerance}
\title{Glucose Tolerance Test Classification}
\usage{
factor_glucose_tolerance(x)
}
\arguments{
\item{x}{A glucose value in mg/dL. Can be single value or a vector. Converts
a number entered as a string to numeric if possible, otherwise returns \code{NA}}
}
\value{
A Factor for OGTT classification of fasting and 2hr glucose
}
\description{
Factor the 2-hour glucose from a standard 75g OGTT
}
\details{
\itemize{
\item \code{factor_glucose_tolerance()} classifies glucose tolerance from a 2-hr OGTT
glucose as Normal (NGT), Impaired (IGT), or Diabetes (DM2)
}
}
\examples{
factor_glucose_tolerance(140)
factor_glucose_tolerance(139.9)
factor_glucose_tolerance(199.9)
factor_glucose_tolerance(200)
factor_glucose_tolerance("0")
}
