% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_glucosetolerance.R
\name{factor_glucosetolerance_0}
\alias{factor_glucosetolerance_0}
\title{Glucose Tolerance Test Classification}
\usage{
factor_glucosetolerance_0(x)
}
\arguments{
\item{x}{A glucose value in mg/dL. Can be single value or a vector. Converts
a number entered as a string to numeric if possible, otherwise returns `NA`}
}
\value{
A Factor for OGTT classification of fasting and 2hr glucose
}
\description{
Functions to factor the fasting and 2-hour glucose from a standard 75g OGTT
}
\details{
* `factor_glucosetolerance_0()` classifies fasting glucose as Normal (NFG),
Impaired (IFG), or Diabetes (DM2)
* `factor_glucosetolerance_120()` classifies glucose tolerance from a 2-hr OGTT
glucose as Normal (NGT), Impaired (IGT), or Diabetes (DM2)
}
\examples{
factor_glucosetolerance_120(140)
factor_glucosetolerance_120(139.9)
factor_glucosetolerance_120(199.9)
factor_glucosetolerance_120(200)
factor_glucosetolerance_120("0")
factor_glucosetolerance_0(c(120, 130, 150, 200))
}
