% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_barmean.R
\name{geom_barmean}
\alias{geom_barmean}
\title{Summary bar plot (with optional error bar, overlying points)}
\usage{
geom_barmean(se = TRUE, points = F, barfill = "steelblue")
}
\arguments{
\item{se}{add SEM bars (TRUE/FALSE)}

\item{points}{add jittered points (uses `ggbeeswarm::geom_quasirandom()`)
(TRUE/FALSE)}

\item{barfill}{color for bar fill (as a string, default = "steelblue")}
}
\value{
a ggplot object
}
\description{
Plots a summary Barplot of the averaged group values. Optionally can add
error bars (SEM) and jittered data points. Does not have full functionality
of `geom_bar()`, e.g., cannot pass additional aes() elements or arguments.
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(factor(cyl), disp))
p +  geom_barmean()
p +  geom_barmean(points = T)
p +  geom_barmean(points = T, barfill = "purple") 
}
