% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_lm.R
\name{geom_lm}
\alias{geom_lm}
\title{linear regression version of geom_smooth()}
\usage{
geom_lm(formula = y ~ x, se = TRUE, linewidth = 1, ...)
}
\arguments{
\item{formula}{}

\item{se}{Plot error range (T/F)}

\item{linewidth}{width of regression line}

\item{...}{pass additional arguments}
}
\value{
a ggplot object
}
\description{
Convenience function- a version of `geom_smooth()` using `lm()` instead of `loess()`
}
\examples{
Orange \%>\% 
  ggplot(aes(age, circumference)) +
    geom_lm(alpha = 0.5, color = "steelblue") +
      geom_point() 
}
