% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_pointline_sem.R
\name{geom_pointline_sem}
\alias{geom_pointline_sem}
\title{ggplot group summary stat with line, point and error bar}
\usage{
geom_pointline_sem(point = T, line = T, pd = 0, fun = "mean", error = T)
}
\arguments{
\item{point}{add points (TRUE/FALSE)}

\item{line}{add connecting lines (TRUE/FALSE)}

\item{pd}{amount to use for dodge (use if multiple groups/lines)}

\item{fun}{summary function used (default = "mean")}

\item{error}{add SEM bars (TRUE/FALSE)}
}
\value{
a ggplot object
}
\description{
ggplot group summary stat with line, point and error bar
}
\examples{
library(ggplot2)
# Tree Growth

# just summary stats
p <- ggplot(Orange, aes(x = age, y = circumference)) + 
  geom_pointline_sem(pd=0) +
  theme_classic()
p 
# add individual data
p + geom_line(aes(x = age, y = circumference, group = Tree), inherit.aes = F, color = "black", alpha =0.3) 

# Indomethacin PK
p <- ggplot(Indometh, aes(x = time, y = conc)) + 
  geom_pointline_sem(pd=0) +
  scale_y_log10()
  p
 p + geom_line(aes(x = time, y = conc, group = Subject), inherit.aes = F, color = "black", alpha =0.3)
}
