% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls_pvalue.R
\name{gls_pvalue}
\alias{gls_pvalue}
\title{Get p-value from an rms::Gls model}
\usage{
gls_pvalue(gls_model)
}
\arguments{
\item{gls_model}{a Gls model obtained from \code{\link[rms]{Gls}}}
}
\value{
A single p-value
}
\description{
Get p-value from an rms::Gls model
}
\seealso{
\code{\link[rms]{Gls}}
}
