% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabletools.R
\name{htmlToPng}
\alias{htmlToPng}
\title{Converts any html into a png file}
\usage{
htmlToPng(html, zoom = 1, rotate = 0)
}
\arguments{
\item{html}{HTML text to convert}

\item{zoom}{Zoomfactor}

\item{rotate}{Rotation 0 - 360}
}
\value{
A file name of the temporary png file created
}
\description{
Converts any html into a png file
}

