% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_ddd.R
\name{lookup_ddd}
\alias{lookup_ddd}
\title{Lookup Defined Daily Dose (DDD) for hypertension medications}
\usage{
lookup_ddd(medication)
}
\arguments{
\item{medication}{Medication name (internally converted to lowercase)}
}
\value{
Reference DDD value for the medication provided
}
\description{
Lookup function for DDD. Provides values only for hypertension related
medications. Medication name includes only the name of the
active ingredient, not the brand name (e.g. "clonidine" not "clonidine
hydrochloride"). The function will internally convert to lowercase for
matching purposes to the DDD list.
}
\examples{
# example data
htn_med_dose
htn_med_dose |>
  mutate(
    ddd_value = lookup_ddd(medication), # example only - not required
    ddd_calc = calculate_ddd(medication, dose)) 
}
