% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogtt_matsuda.R
\name{matsuda}
\alias{matsuda}
\title{Matsuda-Defronzo Index Calculation (from standard 75g OGTT)}
\usage{
matsuda(time, glucose, insulin)
}
\arguments{
\item{time}{a column name (unquoted) indicating time values (in minutes)}

\item{glucose}{a column name (unquoted) storing glucose values (in mg/dL)}

\item{insulin}{a column name (unquoted) storing insulin values (in $\mu$U/mL)}

\item{time_units}{if units are not in "min", can indicate here for unit conversion (options "min" or "hr")}

\item{glucose_units}{if units are not in "mg/dl", can indicate here for unit conversion (options "mg/dl" or "mmol/l")}

\item{insulin_units}{if units are not in "uU/ml", can indicate here for unit conversion (options "uU/ml" or "pmol/l")}
}
\value{
Matsuda index as a single value
}
\description{
This function calculates the Matsuda index using glucose and insulin 
sampled during a standard 75g oral glucose tolerance test.
Results can be checked using the [Matsuda online calculator](http://mmatsuda.diabetes-smc.jp/MIndex.html)
Standard timepoints are 0, 30, 60, 90, and 120 min.
}
\details{
The Matsuda Composite <i>Insulin Sensitivity Index</i> is calculated as:
$$ ISI_{Composite} = \frac{10,000} {\sqrt{G_{0} \cdot Ins_{0} \cdot \bar{G} \cdot \bar{Ins}}}$$

where:

 *  $G_{0}$ is fasting plasma glucose (in mg/dL)
 *  $Ins_{0}$ is fasting plasma Insulin (in $\mu$U/mL)
 *  $\bar{G}$ is mean glucose during the OGTT (as weighted avg = AUC/time)
 *  $\bar{Ins}$ is mean insulin during the OGTT (as weighted avg = AUC/time)  

Note: insulin unit conversion may differ differ depending on assay. 
Insulin (pmol/l) = insulin (uU/ml)*6 

`matsuda_df()` requires a dataframe input with time, glucose, insulin columns. 
`matsuda()` accepts 3 separate vectors for time, glucose, insulin.
}
\examples{
ogtt1 <- data.frame(
  time=c(0, 30, 60, 90, 120),              # minutes
  glucose=c(93, 129, 178, 164, 97),        # mg/dL
  insulin=c(12.8, 30.7, 68.5, 74.1, 44.0)) # uU/mL
  
# dataframe with variable columns appropriately named
matsuda(time =  ogtt1$time,
        glucose = ogtt1$glucose,
        insulin = ogtt1$insulin) # 3.43125
# error if time is out of order
ogtt3 <- data.frame(
  time=c(0, 60, 30, 90, 120),              # minutes
  glucose=c(100, 129, 178, 164, 97),        # mg/dL
  insulin=c(12.8, 100, 68.5, 74.1, 44.0)) # uU/mL

matsuda(time =  ogtt3$time,
        glucose = ogtt3$glucose,
        insulin = ogtt3$insulin) # error

# error if missing values
ogtt4 <- data.frame(
  time=c(0, 60, 30, 90, 120),              # minutes
  glucose=c(100, NA, 178, 164, 97),        # mg/dL
  insulin=c(12.8, 100, 68.5, 74.1, 44.0)) # uU/mL

matsuda(time =  ogtt4$time,
        glucose = ogtt4$glucose,
        insulin = ogtt4$insulin) # error

# Convert units: 
ogtt5 <- data.frame(time = c(0,0.5,1,1.5,2), # time in hours
                    glucose = c(5.167, 7.167, 9.889, 9.111, 5.3889), # glucose in mmol/l
                    insulin = c(76.8,184.2,411,444.6,264)) # insulin in pmol/l

matsuda(time =  ogtt5$time,
        glucose = ogtt5$glucose,
        insulin = ogtt5$insulin,
        time_units = "hr", insulin_units = "pmol/l", glucose_units = "mmol/l")
}
