% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogtt_matsuda_df.R
\name{matsuda_df}
\alias{matsuda_df}
\title{Matsuda-Defronzo Index Calculation (from standard 75g OGTT)}
\usage{
matsuda_df(
  df,
  timeCol = time,
  glucoseCol = glucose,
  insulinCol = insulin,
  time_units = "min",
  glucose_units = "mg/dl",
  insulin_units = "uU/ml"
)
}
\arguments{
\item{df}{A long-format data frame}

\item{timeCol}{a column name (unquoted) indicating time values (in minutes)}

\item{glucoseCol}{a column name (unquoted) storing glucose values (in mg/dL)}

\item{insulinCol}{a column name (unquoted) storing insulin values (in $\mu$U/mL)}

\item{time_units}{if units are not in "min", can indicate here for unit conversion (options "min" or "hr")}

\item{glucose_units}{if units are not in "mg/dl", can indicate here for unit conversion (options "mg/dl" or "mmol/l")}

\item{insulin_units}{if units are not in "uU/ml", can indicate here for unit conversion (options "uU/ml" or "pmol/l")}
}
\value{
Matsuda index as a single value
}
\description{
This function calculates the Matsuda index using glucose and insulin 
sampled during a standard 75g oral glucose tolerance test.
Results can be checked using the [Matsuda online calculator](http://mmatsuda.diabetes-smc.jp/MIndex.html)
Standard timepoints are 0, 30, 60, 90, and 120 min.
}
\details{
The Matsuda Composite <i>Insulin Sensitivity Index</i> is calculated as:
$$ ISI_{Composite} = \frac{10,000} {\sqrt{G_{0} \cdot Ins_{0} \cdot \bar{G} \cdot \bar{Ins}}}$$

where:

 *  $G_{0}$ is fasting plasma glucose (in mg/dL)
 *  $Ins_{0}$ is fasting plasma Insulin (in $\mu$U/mL)
 *  $\bar{G}$ is mean glucose during the OGTT (as weighted avg = AUC/time)
 *  $\bar{Ins}$ is mean insulin during the OGTT (as weighted avg = AUC/time)  

Note: insulin unit conversion may differ differ depending on assay. 
Insulin (pmol/l) = insulin (uU/ml)*6 

`matsuda_df()` requires a dataframe input with time, glucose, insulin columns. 
`matsuda()` accepts 3 separate vectors with time, glucose, insulin.
}
\examples{
# A long-format dataframe with a single subject
ogtt1 <- data.frame(
  time=c(0, 30, 60, 90, 120),              # minutes
  glucose=c(93, 129, 178, 164, 97),        # mg/dL
  insulin=c(12.8, 30.7, 68.5, 74.1, 44.0)) # uU/mL
  
# dataframe with variable columns appropriately named
  matsuda_df(ogtt1) # works because columns are default names
  matsuda_df(ogtt1, time, glucose, insulin)
  
# Need to specify column names if named differently
ogtt2 <- data.frame(
    t=c(0, 30, 60, 90, 120),              # minutes
    gluc=c(93, 129, 178, 164, 97),        # mg/dL
    ins=c(12.8, 30.7, 68.5, 74.1, 44.0)) # uU/mL
matsuda_df(ogtt2)
matsuda_df(ogtt2, t, gluc, ins) # works by positional matching
matsuda_df(ogtt2, timeCol = t, glucoseCol =  gluc, insulinCol =  ins)

# # error if time data is out of order
ogtt3 <- data.frame(
  time=c(0, 60, 30, 90, 120),              # minutes
  glucose=c(100, 129, 178, 164, 97),        # mg/dL
  insulin=c(12.8, 100, 68.5, 74.1, 44.0)) # uU/mL
  
  matsuda_df(ogtt3, time, glucose, insulin)
  
# error if missing values
ogtt4 <- data.frame(
  time=c(0, 60, 30, 90, 120),              # minutes
  glucose=c(100, NA, 178, 164, 97),        # mg/dL
  insulin=c(12.8, 100, 68.5, 74.1, 44.0)) # uU/mL
  
  matsuda_df(ogtt4, time, glucose, insulin)

# dataframe with differently named variable columns
# Convert units:
ogtt5 <- data.frame(time = c(0,0.5,1,1.5,2), # time in hours
                    glucose = c(5.167, 7.167, 9.889, 9.111, 5.3889), # glucose in mmol/l
                    insulin = c(76.8,184.2,411,444.6,264)) # insulin in pmol/l

# you don't have to specify units, if using default values
matsuda_df(ogtt1, time, glucose, insulin,
           insulin_units = "uU/ml", glucose_units = "md/dl") # default values
# if not default, specify units for proper conversinon:
matsuda_df(ogtt5, time, glucose, insulin,
           time_units = "hr", insulin_units = "pmol/l", glucose_units = "mmol/l")

matsuda_df(ogtt5, time_units = "hr", insulin_units = "pmol/l", glucose_units = "mmol/l")
}
