% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_counts.R
\name{merge_counts}
\alias{merge_counts}
\title{Merge specific columns of a count matrix}
\usage{
merge_counts(x, cols, agg_fun)
}
\description{
This function merges columns of a count matrix before statistical analysis with e.g. MAGeCK
}
