% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_summary.R
\name{my_summary}
\alias{my_summary}
\title{Custom Summarization for a variable in a dataframe}
\usage{
my_summary(df, my_var, probs = c(0.025, 0.975), na.rm = T, digits = 2)
}
\arguments{
\item{df}{A dataframe}

\item{my_var}{A variable name}

\item{probs}{For quantile functions, a vector with the desired quantiles...c(0.025, 0.975) default}

\item{na.rm}{a logical value indicating whether NA values should be stripped}

\item{digits}{integer indicating number of decimal places}
}
\value{
A dataframe with named variables for: mean, sd, n, sem, and 95\%CI
}
\description{
Uses \code{dplyr} style summarization to retrun mean, SD, n, SEM, and 95\% CI for a
single continuous variable in a dataframe. Useful for grouped data summaries.
}
\examples{
library(dplyr)
my_summary(mtcars, mpg)
mtcars \%>\% group_by(cyl) \%>\% my_summary(mpg)
mtcars \%>\% group_by(cyl) \%>\% my_summary(mpg, probs = c(0.01, .999), digits = 0)
}
\seealso{
\code{\link[dplyr]{summarise}}, \code{\link[base]{mean}}, \code{\link[base]{sd}},
\code{\link[base]{round}}, \code{\link[stats]{quantile}}
}
