% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_counts.R
\name{normalize_counts}
\alias{normalize_counts}
\title{Apply normalizations to count matrix}
\usage{
normalize_counts(counts, method = "none")
}
\description{
This function normalizes a count matrix before statistical analysis with e.g. MAGeCK
}
