% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ogtt_wide}
\alias{ogtt_wide}
\alias{ogtt_nested}
\title{Oral Glucose Tolerance Datasets}
\format{
\code{ogtt_wide} A dataframe with 184 OGTT results contains glucose and
insulin in wide format with time noted in the variable name (e.g.,
\code{glucose_000}, \code{glucose_030}). \code{ogtt_nested} contains time, glucose, and
insulin in long format a dataframe stored in a list-column for each
participant.
Demographic data includes:
\describe{
\item{age}{Age, years}
\item{gender.f}{Gender, factored as Male/Female}
\item{race.f}{Race, factored using self-reported classification]}
\item{height_cm}{Height, cm}
\item{weight_kg}{Weight, kg}
\item{bmi}{Body mass index (BMI, \eqn{kg/m^2})}
\item{\code{hdl_bl}, \code{dld_bl}, \code{trigs_bl}}{Fasting lipids: LDL and HDL cholesterol, Triglycerides (mg/dL)}
\item{\verb{glucose_***}}{Plasma Glucose (mg/dL), with time in minutes}
\item{\verb{insulin_***}}{Plasma Insulin (uU/mL), with time in minutes}
\item{\code{ogtt_df}}{A nested dataframe with a single OGTT result per row in the nested dataset (\code{ogtt_nested}), Time points, in minutes (0, 30, 60, 90, 120) denoted either in the variable name (\verb{_000}, \verb{_030}) or in the time column of the nested dataset}
}

\code{ogtt_wide} A dataframe with 184 OGTT results, with a nested
dataframe with each participant result (time, glucose, insulin)
}
\source{
Primary data collected from multiple studies
}
\usage{
ogtt_wide

ogtt_nested
}
\description{
Data obects containing results from stanadard 75g OGTT including
glucose, insulin and demographic data.

All OGTT studies were done after an overnight fast, and all glucose results
are plasma glucose (mg/dL) measured on a YSI instrumement (glucose oxidase).
Plasma Insulin (mU/L or uU/mL). All OGTTs were dones prior to any
intervention (ie, participants were not on study medications).
}
\keyword{datasets}
