% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecg.R
\name{plot_ecg}
\alias{plot_ecg}
\title{Read and Plot Apple Health ECG data in R}
\usage{
plot_ecg(ecg_file, width_max = 125)
}
\arguments{
\item{ecg_file}{ECG data file in csv format}

\item{width_max}{Desired width per row}
}
\value{
ECG as a ggplot
}
\description{
Uses ECG data exported from apple watch ECG, and plots using voltage. Data is converted to standard ECG coordinates internally matching ECG convention (x units = 40ms/box, y units = 0.1mV/box). Data is plotted across 6 rows by default, assuming a 30s tracing, but can be adjusted using the \code{width_max} option.
}
\examples{
library(ggplot2, data.table)
ecg_example_filepath =  system.file("extdata", "ecg_example.csv", package = "reatabletools")
plot_ecg(ecg_example_filepath)
plot_ecg(ecg_example_filepath, width_max = 100)
plot_ecg(ecg_example_filepath, width_max = 125)
plot_ecg(ecg_example_filepath, width_max = 150)
plot_ecg(ecg_example_filepath, width_max = 300)

}
