% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefilter_counts.R
\name{prefilter_counts}
\alias{prefilter_counts}
\title{Filter a count matrix by a minimum value in a selected set of columns}
\usage{
prefilter_counts(file, cols, min_count = 50, id_col = "id")
}
\description{
This function pre-filters a count matrix for statistical analysis with e.g. MAGeCK
}
