% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcorr_clean.R
\name{rcorr_clean}
\alias{rcorr_clean}
\title{Modify rcorr results}
\usage{
rcorr_clean(rcmat, lookup_varTable = NULL)
}
\arguments{
\item{rcmat}{An \code{rcorr} class object created by the function `Hmisc::rcorr``}

\item{lookup_varTable}{A lookup table structured as a dataframe with variable lables.
This is optional and by default is not used.}
}
\value{
An rcorr array with the following matrices:
\itemize{
\item correlation (\verb{$r}). unchanged from the original
\item n (\verb{$n}). unchanged from the original
\item P-value (\verb{$P}). unchanged from the original
\item Simplified P-value (\verb{$P_simplified}). a text matrix with P-values
\item Labeled P-value (\code{P_lab}). a text matrix with significance of P-values indicated with \code{*}, \verb{**}, and \verb{***}. Results are unadjusted.
\item label (\verb{$label}). a text matrix with the \code{r} and \code{p-value} summary, useful as a hovertext
\item label (\verb{$label_html}). a text matrix with the \code{r} and \code{p-value} summary, useful to pass to \code{htmlTable}
}
}
\description{
Modifies the \code{rcorr} output from the \code{Hmisc::rcorr()} function.
The result is an array with additional matrices (\verb{$P_simplified}, \verb{$label}, and \code{P_lab}).
Optionally will replace the variable names with labels from a lookup table (see example)
}
\details{
Modify the Hmisc::rcorr output for nicer printing with labels in a heatmap
}
\examples{
require(Hmisc)
require(tibble)
require(heatmaply)
require(ggplot2)
# create a LOOKUP TABLE
lookup_varTable <-
  tibble::tribble(~varname, ~labels,
          "mpg",	"Miles/(US) gallon",
          "cyl",	"Number of cylinders",
          "disp",	"Displacement (in.<sup>3</sup>)",
          "hp",	  "Gross horsepower",
          "drat",	"Rear axle ratio",
          "wt",	  "Weight (1000 lbs)",
          "qsec",	"1/4 mile time",
          "vs",	  "Engine (0 = V-shaped, 1 = straight)",
          "am",	  "Transmission (0 = automatic, 1 = manual)",
          "gear",	"Number of forward gears",
          "carb",	"Number of carburetors"  )
test <- Hmisc::rcorr(as.matrix(mtcars) )
test_result <- rcorr_clean(test, lookup_varTable = lookup_varTable)

# make an interactive HEATMAP with heatmaply
heatmaply::heatmaply(test_result$r,
                     custom_hovertext = test_result$label,
                     cellnote = test_result$P_lab,
                     na.value = "black",
                     scale_fill_gradient_fun = ggplot2::scale_fill_gradient2(low = "blue", high = "red", midpoint = 0, limits = c(-1, 1)),
                     cellnote_textposition = "bottom center",
                     key.title = "Spearman's rho:")
}
\seealso{
\code{\link[Hmisc]{rcorr}}, \code{\link[heatmaply]{heatmaply}}
}
