% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdi_decile.R
\name{sdi_decile}
\alias{sdi_decile}
\title{Get Social Deprivation Index (SDI) by Zipcode}
\usage{
sdi_decile(zipcode, year = 2019)
}
\arguments{
\item{zipcode}{Zipcode area}

\item{year}{Year (default to most recent = 2019)}
}
\value{
Single value representing the SDI Decile (lower is better)
}
\description{
Return the SDI decile calculated by 5-year intervals. SDI decile estimates
are available for years ending in 2012, 2015, 2016, 2017, 2018, 2019.
Quantiles are calculated by the SDI score for each year. Detailed description
and source data from
\href{https://www.graham-center.org/maps-data-tools/social-deprivation-index.html}{Robert
Graham Center}.
}
\examples{
sdi_decile(37220)
sdi_decile(90210)
}
