% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabletools-package.R
\docType{package}
\name{tabletools-package}
\alias{tabletools}
\alias{tabletools-package}
\title{tabletools: Helpful Tools to Make html Tables}
\description{
Functions for creating html tables, and other misc. functions.
}
\keyword{internal}
