% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ecg.R
\name{theme_ecg}
\alias{theme_ecg}
\title{ECG ggplot theme
ggplot theme for use in plotting Apple health ECG data in R.}
\usage{
theme_ecg()
}
\description{
ECG ggplot theme
ggplot theme for use in plotting Apple health ECG data in R.
}
