% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jml.R
\name{theme_jml}
\alias{theme_jml}
\title{my preferred basic ggplot theme
R code for my preferred ggplot theme}
\usage{
theme_jml(
  base_size = 11,
  base_family = "sans",
  labels_color = "black",
  axis_color = "black",
  bg_color = "transparent",
  grid_color = "grey90",
  legend.position = "bottom"
)
}
\description{
Can be called as a function from this package, or can be loaded from a Github gist containing the same code.
}
\examples{
library(tidyverse)
pl_mtc <- mtcars |> 
 ggplot(aes(cyl, mpg)) +
 geom_point()
 
 pl_mtc + theme_gray()
 pl_mtc + theme_classic()
 pl_mtc + theme_jml()

#OGTT example:
data("ogtt_nested")
ogtt_long <- 
 ogtt_nested |> 
 select(id, ogtt_df) |> 
 unnest(cols = ogtt_df) |> 
 group_by(id)

pl_ogtt <- ogtt_long |> 
 ggplot(aes(time, glucose, group=id)) +
 geom_line(alpha=0.3)
 
 pl_ogtt + theme_jml()
 
#' to make this the active theme for all plots:
theme_set(theme_jml())
pl_ogtt
}
