% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt_chisquare_report.R
\name{txt_chisquare_report}
\alias{txt_chisquare_report}
\title{Summary text functions for html tables}
\usage{
txt_chisquare_report(chisquare_object)
}
\arguments{
\item{chisquare_object}{A Chi-square object}
}
\description{
This function returns html formatted text summary of a Chi-sqauare test. 
The input object must be a chisq.test object, and the p-value is formatted by the 
`htmlTable::txtPval()` function.
}
\examples{
library(htmlTable)
# a basic 2x2 table
table(mtcars$vs,mtcars$cyl) # the table
cstest1 <-chisq.test(table(mtcars$vs,mtcars$am)) # the chi.square test object
htmlTable::htmlTable(txt_chisquare_report(cstest1))

# a 2x3 table
table(mtcars$vs,mtcars$cyl)
cstest2 <-chisq.test(table(mtcars$vs,mtcars$cyl))
htmlTable::htmlTable(txt_chisquare_report(cstest2))
}
\seealso{
\code{\link[htmlTable]{txtPval}}, \code{\link[base]{sprintf}}, \code{\link[stats]{chisq.test}}
}
