% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_cleaning_functions.R
\name{txt_clean_nonascii}
\alias{txt_clean_nonascii}
\title{Functions to clean character strings}
\usage{
txt_clean_nonascii(string, regex = "[^[:alnum:][:blank:]+?&/\\\\-\\\\.]")
}
\arguments{
\item{string}{character string, or vector of character strings which will be cleaned}

\item{regex}{defaults to remove all non-ascii text characters, but keeping a few}
}
\value{
A character string or vector, now stripped of any non-matching values
}
\description{
\code{txt_clean_nonascii} removes non-standard characters from text, but keeps space, -, and . by default, using regex matching.
Default matching can be overridden by \code{regex} option.
}
\examples{
test_strings <- c("\"The Erlenmeyer Flask\"‡", "\"My Struggle\"‡", "\"Founder's Mutation\"", "\"Mulder & Scully Meet the Were-Monster\"",
"\"Home Again\"", "\"Babylon\"", "\"My Struggle II\"‡" )
test_strings
txt_clean_nonascii(test_strings)
}
