% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt_mcnemar_report.R
\name{txt_mcnemar_report}
\alias{txt_mcnemar_report}
\title{Summary text functions for html tables}
\usage{
txt_mcnemar_report(mcnemar_object)
}
\arguments{
\item{mcnemar_object}{A McNemar's Chi-square object}
}
\description{
This function returns html formatted text summary of McNemar's Chi-sqauare test, for paired data.
The input object must be a \code{mcnemar.test} object, and the p-value is formatted by the
\code{htmlTable::txtPval()} function.
}
\examples{
ex <- data.frame(A = sample(c("No", "Yes"), 9, replace = T),
                 B = sample(c("No", "Yes"), 9, replace = T))
ex
result <- mcnemar.test(table(ex$A, ex$B))
txt_mcnemar_report(result)
}
\seealso{
\code{\link[htmlTable]{txtPval}}, \code{\link[base]{sprintf}}, \code{\link[stats]{mcnemar.test}}
}
