% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt_mean_sd.R
\name{txt_mean_sd}
\alias{txt_mean_sd}
\title{Summary text functions for html tables}
\usage{
txt_mean_sd(var, na.rm = TRUE, ...)
}
\arguments{
\item{var}{A numeric vector or variable name within a dataframe}

\item{na.rm}{a logical value indicating whether NA values should be stripped}

\item{...}{Values to be passed into `mean` arguments.}
}
\value{
A character vector with mean and sd, suitable for printing in an html table
}
\description{
Summary text functions for html tables
}
\examples{
library(dplyr)
txt_mean_sd(mtcars$mpg)
txt_mean_sd(mtcars$mpg, trim = .2)
txt_mean_sem(mtcars$mpg)
txt_mean_range(mtcars$mpg)
txt_median_iqr(mtcars$mpg)
txt_pct_fr(mtcars$cyl, 6)
mtcars \%>\%
  group_by(cyl) \%>\%
  summarise(mean_mpg = txt_mean_sd(mpg))
}
\seealso{
\code{\link[base]{sprintf}}, \code{\link[base]{mean}}, \code{\link[base]{sd}}
}
