% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt_mean_sem.R
\name{txt_mean_sem}
\alias{txt_mean_sem}
\title{Summary text functions for html tables}
\usage{
txt_mean_sem(var, na.rm = TRUE, digits = 2)
}
\arguments{
\item{var}{A numeric vector or variable name within a dataframe}

\item{na.rm}{a logical value indicating whether NA values should be stripped}

\item{digits}{number of decimal places (limited to 1-3)}
}
\value{
A character vector with mean and sem, suitable for printing in an html table
}
\description{
Summary text functions for html tables
}
\examples{
txt_mean_sem(mtcars$mpg)
txt_mean_sem(mtcars$mpg, digits=1)
}
\seealso{
\code{\link[base]{sprintf}}, \code{\link[base]{mean}}, \code{\link[base]{sd}}
}
