% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt_median_iqr.R
\name{txt_median_iqr}
\alias{txt_median_iqr}
\title{Summary text functions for html tables}
\usage{
txt_median_iqr(var, lower = 0.25, upper = 0.75, na.rm = T, ...)
}
\arguments{
\item{var}{A vector, or variable name within a dataframe}

\item{lower}{For quantile functions, the lower quantile (0.25 default)}

\item{upper}{For quantile functions, the upper quantile (0.75 default)}

\item{na.rm}{a logical value indicating whether NA values should be stripped}

\item{...}{passes to `median(...)`}
}
\value{
A character vector with median (interquartile range), suitable for printing in an html table
}
\description{
Summary text functions for html tables
}
\examples{
library(dplyr)
txt_median_iqr(mtcars$mpg)
mtcars \%>\%
  group_by(cyl) \%>\%
  summarise(mean_mpg = txt_mean_range(mpg))
}
\seealso{
\code{\link[base]{sprintf}}, \code{\link[stats]{median}}, \code{\link[base]{quantile}}
}
