% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt_pct_fr.R
\name{txt_pct_fr}
\alias{txt_pct_fr}
\title{Summary text functions for html tables}
\usage{
txt_pct_fr(var, referent_value, na.rm = TRUE)
}
\arguments{
\item{var}{A vector, or variable name within a dataframe}

\item{referent_value}{A character string to define the reference value}

\item{na.rm}{a logical value indicating whether NA values should be stripped}
}
\description{
Summary text functions for html tables
}
\examples{
library(dplyr)
txt_pct_fr(mtcars$gear, 4)
mtcars \%>\%
  group_by(cyl) \%>\%
  summarise(pct_4gears = txt_pct_fr(gear, 4)) # shows what percentage have 4 gears (selected reference) and (with fraction)
}
