% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{bmi}
\alias{bmi}
\title{Body Mass Index Calculation}
\usage{
bmi(weight, height, weight_units = "kg", height_units = "m", ...)
}
\arguments{
\item{weight}{Body Weight (kg)}

\item{height}{Body Height (m)}

\item{weight_units}{If `weight` is in Units other than kg, specifiy here to  properly convert for calucations ("kg", "g", or "lbs")}

\item{height_units}{If `height` is in Units other than m, specifiy here to  properly convert for calucations ("m", "cm", or "in")}
}
\value{
a numeric vector with Body Mass Index (BMI, kg/m^2)
}
\description{
Calculation for body mass index (BMI, in kg/m^2), with unit conversion if needed.
}
\examples{
bmi(weight = 70, height = 1.778)
# different units
bmi(weight = 154, weight_units = "lbs", 
    height = 70, height_units = "in", 
    method = "Mosteller")
bmi(weight = 70, height = 1.778)

}
