% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsa.R
\name{bsa}
\alias{bsa}
\title{Body Surface Area Calculation}
\usage{
bsa(
  weight,
  height,
  weight_units = "kg",
  height_units = "m",
  method = "Mosteller",
  ...
)
}
\arguments{
\item{weight}{Body Weight (kg)}

\item{height}{Body Height (m)}

\item{weight_units}{If `weight` is in Units other than kg, specifiy here to  properly convert for calucations ("kg", "g", or "lbs")}

\item{height_units}{If `height` is in Units other than m, specifiy here to  properly convert for calucations ("m", "cm", or "in")}

\item{method}{Which BSA formula to use for calculation ()}
}
\value{
a numeric vector
}
\description{
Calculates Body Surface Area (m2) by several common methods. Converts weight to Kg and Height to cm if needed. 

Avaialbe methods for BSA calculations include:
\deqn{\sqrt{\frac{Ht(cm) \cdot Wt(kg)}{3600}} \tag{Mosteller}}

\deqn{Wt(kg)^{0.425} \cdot Ht(cm)^{0.725} \cdot 0.007184 \tag{DuBois and DuBois}}

\deqn{Wt(kg)^{0.51456} \cdot Ht(cm)^{0.42246} \cdot 0.0235 \tag{Gehan and George}}

\deqn{Wt(kg)^{0.5378} \cdot Ht(cm)^{0.3964} \cdot 0.024265 \tag{Haycock}}

\deqn{71.3989 \cdot Ht(cm)^{0.7437} \cdot Wt(kg)^{0.4040} \div 10000 \tag{Yu}}

\deqn{0.1173 \cdot Wt(kg)^{0.6466} \tag{Livingston}}

\deqn{128.1 \cdot Ht(cm)^{0.60} \cdot Wt(kg)^{0.44} \tag{Tikuisis}}
}
\examples{
bsa_methods <- c("Mosteller", "DuBois", "Boyd", "Gehan-George", "Haycock", "Yu", "Livingston", "Tikuisis")
names(bsa_methods) <- bsa_methods
sapply(bsa_methods, \(x)bsa(weight = 70, height = 1.778, method = x))
bsa(weight = 70, height = 1.778)
bsa(weight = 70, height = 1.778, method = "Mosteller")
bsa(weight = 70, height = 1.778, method = "DuBois")
bsa(weight = 70, height = 1.778, method = "Boyd")
bsa(weight = 70, height = 1.778, method = "Gehan-George")
bsa(weight = 70, height = 1.778, method = "Haycock")
bsa(weight = 70, height = 1.778, method = "Yu")
bsa(weight = 70, height = 1.778, method = "Livingston")
bsa(weight = 70, height = 1.778, method = "Tikuisis")

# different units
bsa(weight = 154, weight_units = "lbs", 
    height = 70, height_units = "in", 
    method = "Mosteller")
bsa(weight = 70, height = 1.778)
}
