% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_creatclearance_cg.R
\name{calculate_creatclearance_cg}
\alias{calculate_creatclearance_cg}
\title{Calculate estimated GFR by Cockcroft-Gault}
\usage{
calculate_creatclearance_cg(age, sex, weight, creatinine, weight_units = "kg")
}
\arguments{
\item{age}{Age, in years}

\item{sex}{Sex}

\item{weight}{Weight, in kg}

\item{creatinine}{Serum creatinine, in mg/dL}
}
\value{
a numeric vector with eGFR (ml/min/1.73m^2)
}
\description{
Estimates Creatinine Clearance based on age, sex, weight, and
creatinine. Note that other estimating equations estimate GFR.
}
\examples{
calculate_creatclearance_cg(age=70, sex="MAle", weight=70, creatinine=1.0) # 68.05556
calculate_creatclearance_cg(age=70, sex="ma", weight=70, creatinine=1.0) # 57.84722
calculate_creatclearance_cg(age=50, sex="Male", weight=70, creatinine=1.0) # 87.5
calculate_creatclearance_cg(age=50, sex="F", weight=70, creatinine=1.0) # 74.375
calculate_creatclearance_cg(age=50, sex="fe", weight=154, creatinine=1.0, weight_units="lbs") # 74.375
}
