% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_egfr.R, R/calculate_egfr_ckdepi.R
\name{calculate_egfr_ckdepi}
\alias{calculate_egfr_ckdepi}
\title{Calculate estimated GFR by CKD-EPI equation}
\usage{
calculate_egfr_ckdepi(age, sex, creatinine)

calculate_egfr_ckdepi(age, sex, creatinine)
}
\arguments{
\item{age}{Age, in years}

\item{sex}{Sex}

\item{creatinine}{Serum creatinine in mg/dL}
}
\value{
a numeric vector with eGFR (ml/min/1.73m^2)

a numeric vector with eGFR (ml/min/1.73m^2)
}
\description{
Calculates estimated glomerular filtration rate (GFR) by several
  common methods. Converts weight to Kg and Height to cm if needed. Avaialbe
  methods for eGFR calculations include:

  *  CKD-EPI(2021): 142*(Cr/A)^B * 0.9938^age * (1.012 if Female)
  *  CKD-EPI Cystatin-C: 135 x (Scr/A)B x (Scys/C)D x 0.9961age x (0.963 if female)
  *  MDRD:
  *  Cockcroft-Gault:

Calculates estimated glomerular filtration rate (GFR) by several
  common methods. Converts weight to Kg and Height to cm if needed. Avaialbe
  methods for eGFR calculations include:

  *  CKD-EPI(2021): 142*(Cr/A)^B * 0.9938^age * (1.012 if Female)
  *  CKD-EPI Cystatin-C: 135 x (Scr/A)B x (Scys/C)D x 0.9961age x (0.963 if female)
  *  MDRD:
  *  Cockcroft-Gault:
}
\examples{
calculate_egfr(age=70, sex="Male", creatinine=0.8, method="CKD-EPI") # 95.2
calculate_egfr(age=50, sex="Female", creatinine=1.0, method="CKD-EPI") # 68.6  
calculate_egfr(age=70, sex="Male", creatinine=0.8, method="CKD-EPI") # 95.2
calculate_egfr(age=50, sex="Female", creatinine=1.0, method="CKD-EPI") # 68.6  
}
