% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_egfr.R, R/calculate_egfr_ckdepi.R
\name{calculate_egfr_ckdepi}
\alias{calculate_egfr_ckdepi}
\alias{calculate_egfr_ckdepi_nonv}
\title{Calculate estimated GFR by CKD-EPI equation}
\usage{
calculate_egfr_ckdepi(age, sex, creatinine, race = NULL, version = "2021")

calculate_egfr_ckdepi_nonv(age, sex, creatinine, race = NULL, version = "2021")
}
\arguments{
\item{age}{Age, in years}

\item{sex}{Sex}

\item{creatinine}{Serum Creatinine, mg/dL}

\item{race}{Race limited to either "White" or "Black"}

\item{version}{Equation version, either "2021" or "original"}
}
\value{
a numeric vector with eGFR (ml/min/1.73m^2)

a numeric vector with eGFR (ml/min/1.73m^2)
}
\description{
Calculates estimated glomerular filtration rate (GFR) by several
  common methods. Converts weight to Kg and Height to cm if needed. Avaialbe
  methods for eGFR calculations include:

  *  CKD-EPI(2021): 142*(Cr/A)^B * 0.9938^age * (1.012 if Female)
  *  CKD-EPI Cystatin-C: 135 x (Scr/A)B x (Scys/C)D x 0.9961age x (0.963 if female)
  *  MDRD:
  *  Cockcroft-Gault:

Calculates estimated glomerular filtration rate (GFR) by several
  common methods. Converts weight to Kg and Height to cm if needed. Available
  methods for eGFR calculations include:

  * CKD-EPI AS Equation (2021) incorporating only Age, Sex 
  * CKD-EPI ASR Equation (original, 2009) incorporating Age, Sex, Race
  * CKD-EPI Cystatin-C (not incorporated)
  * MDRD  
  * Cockcroft-Gault 

Online calculators can be used to compare reults for \href{https://www.kidney.org/professionals/kdoqi/gfr_calculator}{CKD-EPI AS (2021)},
\href{https://www.niddk.nih.gov/health-information/professionals/clinical-tools-patient-management/kidney-disease/laboratory-evaluation/glomerular-filtration-rate-calculators/historical}{CKD-EPI ASR}
}
\examples{
calculate_egfr(age=70, sex="Male", creatinine=0.8, method="CKD-EPI") # 95.2
calculate_egfr(age=50, sex="Female", creatinine=1.0, method="CKD-EPI") # 68.6  
# CKD-EPI 2021 version (new race-free creatinine-based equation)
# https://www.kidney.org/professionals/kdoqi/gfr_calculator
calculate_egfr_ckdepi(age=50, sex="Male", creatinine=0.6) # 
calculate_egfr_ckdepi(age=70, sex="Male", creatinine=0.8, version = "2021") # 95.2
calculate_egfr_ckdepi(age=70, sex="Female", creatinine=0.8, version = "2021") # 79.2

# use of race-incorporated version not recommended (ASR);
# if you want to use race, use version = "original".
calculate_egfr_ckdepi(age=70, sex="Female", creatinine=0.8, race="White", version = "2021") # error
calculate_egfr_ckdepi(age=70, sex="Female", creatinine=0.8, race="White", version = "original") # 74

# original CKD-EPI equation; see [NIH eGFR calculator](https://www.niddk.nih.gov/health-information/professionals/clinical-tools-patient-management/kidney-disease/laboratory-evaluation/glomerular-filtration-rate-calculators/historical)
calculate_egfr_ckdepi(age=70, sex="Male", creatinine=0.8, version = "original", race = "Black") # 104.2
calculate_egfr_ckdepi(age=70, sex="Male", creatinine=0.8, version = "original", race = "White") # 89.9

# CKD-EPI AS (2021) RESULTS; Race-free equation
# Note: do not use race in function call; results same regardless of race
# results verified against Table S11 in Inker NEJM 2021: eGFR (AS) new
df <-  data.frame(age=rep(c(rep(50,4),rep(75,4)),6),
                  sex = rep(c(rep("Male",8),rep("Female",8)),3),
                  race=c(rep("White",16),rep("Black",16),rep("White",16)),
                  creatinine = rep(c(0.6, 1, 1.5, 2), 6),
                  version = c(rep("2021", 16),rep("original", 32)))
df |>
  mutate(egfr = calculate_egfr_ckdepi(age=age, sex=sex, race=race, creatinine=creatinine, version=version),
         egfr_ckepi_original = calculate_egfr_ckdepi(age=age, sex=sex, race=race, creatinine=creatinine, version="original")) 
}
