% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_egfr_mdrd.R
\name{calculate_egfr_mdrd}
\alias{calculate_egfr_mdrd}
\title{Calculate estimated GFR by MDRD}
\usage{
calculate_egfr_mdrd(age, sex = "Male", race = "White", weight, creatinine)
}
\arguments{
\item{age}{Age, in years}

\item{sex}{Sex}

\item{race}{}

\item{weight}{}

\item{creatinine}{}
}
\value{
a numeric vector with eGFR (ml/min/1.73m^2)
}
\description{
Calculates estimated glomerular filtration rate (GFR) by several
  common methods. Converts weight to Kg and Height to cm if needed. Available
  methods for eGFR calculations include:

  *  CKD-EPI(2021)
  *  CKD-EPI Cystatin-C
  *  MDRD
  *  Cockcroft-Gault
}
\examples{
calculate_egfr_mdrd(age = 50, sex = "Male", race = "White", weight = 70, creatinine = 1.5) # 49.5
calculate_egfr_mdrd(age = 50, sex = "Male", race = "Black", weight = 70, creatinine = 1.5) # 60
calculate_egfr_mdrd(age = 50, sex = "Female", race = "White", weight = 70, creatinine = 1.5) # 36.8
calculate_egfr_mdrd(age = 50, sex = "Female", race = "Black", weight = 70, creatinine = 1.5) # 44.5
}
