% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ibw.R
\name{calculate_ibw}
\alias{calculate_ibw}
\title{Ideal Body Weight Calculation}
\usage{
calculate_ibw(
  height,
  gender = NA,
  weight_units = "lbs",
  height_units = "in",
  method = "Hamwi"
)
}
\arguments{
\item{height}{Body Height (m)}

\item{weight_units}{Desired units for result}

\item{height_units}{If \code{height} is in Units other than m, specifiy here to
properly convert for calucations ("m", "cm", or "in")}

\item{method}{string for IBW formula to use for calculation ("Hamwi",
"Devine", "Robinson", "Broca", "Miller", "Hammond")}
}
\value{
a numeric vector with Ideal Body Weight (defaults to lbs)
}
\description{
Calculates Ideal Body Weight (IBW) by several common methods.
Converts weight to Kg and Height to cm if needed. Available methods for IBW
calculations are described in
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4841935}{Peterson et
al.}:
}
\examples{
ibw_methods <- c("Hamwi", "Devine", "Robinson", "Broca", "Miller", "Hammond")
names(ibw_methods) <- ibw_methods
sapply(ibw_methods, \(x)calculate_ibw(height = 70, gender="M", method = x))
sapply(ibw_methods, \(x)calculate_ibw(height = 70, gender="F", method = x))
# if gender missing returns NA
sapply(ibw_methods, \(x)calculate_ibw(height = 70,  method = x))

# Change Units for desired IBW result
# if not specified defaults to LBS
sapply(ibw_methods, \(x)calculate_ibw(height = 70, gender="F", method = x))
sapply(ibw_methods, \(x)calculate_ibw(height = 70, gender="F", method = x, weight_units="kg"))
sapply(ibw_methods, \(x)calculate_ibw(height = 70, gender="F", method = x, weight_units="kilograms"))
sapply(ibw_methods, \(x)calculate_ibw(height = 70, gender="F", method = x, weight_units="g"))
sapply(ibw_methods, \(x)calculate_ibw(height = 70, gender="F", method = x, weight_units="lbs"))
}
