% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lvmass.R
\name{calculate_lvmass}
\alias{calculate_lvmass}
\title{Calculate LV Mass using Echocardiogram results}
\usage{
calculate_lvmass(lvedd, ivsd, lvpwd, bsa, length_units = "cm")
}
\arguments{
\item{lvedd}{LV Internal Diameter Diastole}

\item{ivsd}{LV Septal Thickness Diastole}

\item{lvpwd}{LV Posterior Thickness Diastole}

\item{bsa}{Body surface area (\eqn{m^2})}

\item{length_units}{units of length used for length (default cm)}
}
\value{
dataframe of echocardiagraphic LV mass results (LVM, LVMI, RWT)
}
\description{
<<<<<<< HEAD
Use commonly reported echocardiographic measurements to calculate Left Ventricular Mass (LVM).
}
\details{
Uses methods according to \href{https://pubmed.ncbi.nlm.nih.gov/25559473/}{American Society of Echocardiography}.
Calculations using formulas, with units converted to mm for calculation:
=======
Uses methods according to \href{https://pubmed.ncbi.nlm.nih.gov/25559473/}{American Society of Echocardiography}.
Calculations using formulas, with units converted to mm for calculation:
}
\details{
>>>>>>> 71f82c244af5fcf80a3e2f59668663afbe9803fd
\itemize{
\item LV Mass (LVM) = \eqn{0.8 \cdot (1.04 \cdot ((LVEDD + IVSD + LVPWD)^3 - LVEDD^3)) + 0.6}
\item LV Mass Index (LVMI) = \eqn{LVM/BSA}
\item Relative Wall Thickness (RWT) = \eqn{2 \cdot LVPWD/LVEDD}
}
}
\examples{
calculate_lvmass(lvedd = 50.5, ivsd = 10.7, lvpwd = 13.3, bsa=2.42, length_units = "mm")
}
