% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_map.R
\name{calculate_map}
\alias{calculate_map}
\title{Calculate Mean Arterial Pressure}
\usage{
calculate_map(sbp, dbp, hr = NA, method = "gomez")
}
\arguments{
\item{sbp}{Systolic blood pressure, mmHg}

\item{dbp}{Diastolic blood pressure, mmHg}

\item{hr}{Heart rate, bpm}

\item{method}{MAP formula, one of `gomez`, `standard`, or `razminia`. Uses
"gomez" method by default. If no HR is supplied, the "standard" method is
used. If `standard` method is used without specifying it, due to no HR, a
warning is issued.}
}
\value{
map as a single value, or vector of results
}
\description{
Returns the mean arterial pressure (MAP) from standard systolic (SBP) and
diastolic (DBP) blood pressure measurements. Note it is best to record MAP
reported directly by the device. Heart rate adjusted methods reported by
\href{https://pubmed.ncbi.nlm.nih.gov/1860182/}{Gomez} and by
\href{https://pubmed.ncbi.nlm.nih.gov/15558774/}{Razminia} are also
supported. The Gomez method is applied as the default, for use in calculation
in renal glomerular pressure estimates.
}
\examples{
calculate_map(159, 46)
calculate_map("159", 46)
calculate_map("159", NA)
calculate_map(140, 90)
calculate_map(140, 90, 70, method="standard")
calculate_map(140, 90, 70, method = "gomez")
calculate_map(140, 90, 70, method="razminia")

plot(hr, map_dbl(hr, ~calculate_map(125, 73, .x, method = "gomez")), "l", lty=1, col="black",
     # xlim = c(30,150), ylim = c(85,100),
     ylab = "Calculated MAP (mmHg)")
title("Methods comparison for HR-adjusted MAP calculation")
lines(hr, map_dbl(hr, ~calculate_map(125, 73, .x, method = "razminia")), lty=2, col="red")
abline(h=calculate_map(125,73, method="standard"),lty=2, col="blue")
}
