% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mgfr_msp.R
\name{calculate_mgfr_msp}
\alias{calculate_mgfr_msp}
\title{Iohexol Multiple Sample Protocol (MSP) estimate of GFR}
\usage{
calculate_mgfr_msp(
  time,
  iohexol_conc,
  omnipaque_v = 300,
  ioh_inj_vol = 5,
  ioh_inj_wt = NULL,
  ioh_units = "ug/mL",
  time_units = "min",
  id = NULL,
  height = NA,
  height_units = "m",
  weight = NA,
  weight_units = "kg",
  method_adj = "BM",
  t_late = 120,
  output = "summary"
)
}
\arguments{
\item{time}{A vector of time values (minutes)}

\item{iohexol_conc}{A vector of Iohexol plasma measurements (ug/mL)}

\item{omnipaque_v}{Omnipaque version (eg 300 = Omnipaque 300 recommended).
Other versions commonly available include 300 and 350 (USA) and 140, 280,
240 (other areas)}

\item{ioh_inj_vol}{Iohexol injection Volume by syringe volume injected (Not
preferred; use if weights not available)}

\item{ioh_inj_wt}{Iohexol injection Weight by syringe weight determination
(Pre-Post weight difference; recommnded method) in grams}

\item{ioh_units}{Iohexol concentration units, defaults to \code{ug/mL}}

\item{time_units}{Time units, defaults to \code{min}}

\item{id}{Study identifier (optional, passed to plot title)}

\item{height}{Patient Height, m}

\item{height_units}{Height units, if not in m}

\item{weight}{Patient Weight, kg}

\item{weight_units}{Weight units, if not in kg}

\item{t_late}{First Time point (minutes) to use in the "Late" elimination
phase, defaults to \code{120}}

\item{output}{Desired output, defaults to \code{summary} of model. Alternatively
can specify \code{gfr}, \code{gfr_bsa}, \code{fit}, or \code{plot}

The \code{mgfr_method} describes the method for estimation, which can have the
following options:
\itemize{
\item \code{SI} Slope-Intercept method obtained by performing two linear regressions of log(iohexol) vs time for early and late time points. Obtained by setting \code{nls_v = "SI"}
\item \code{modified-SI} SI method with shared time-point(s) between early and late periods (ie overlapping \code{t_early} = \code{t_late}). Obtained by setting, for example \verb{nls_v = "SI", t_early=120, t_late=120} or another option with overlapping times.
\item \code{NLLS-gslnls-weighted} (Default method) NLS fit using the \code{gslnls()} function from the \code{gsl_nls} package. Obtained by setting \verb{nls_v = "gslnls", nls_weights=T}
\item \code{NLLS-base-weighted} NLS fit using the \code{nls()} function from the base \code{stats} package. Obtained by setting \verb{nls_v = "base", nls_weights=T}. Similar to \code{gslnls} method, except that parameter constraints cannot be set with \code{nls()}, and so implausible estimates (negative for a, b) may be obtained.
\item \code{NLLS-gslnls-unweighted} NLS fit using the \code{gslnls()} function from the \code{gsl_nls} package. Obtained by setting \verb{nls_v = "gslnls", nls_weights=F}
\item \code{NLLS-base-unweighted} NLS fit using the \code{nls()} function from the base \code{stats} package. Obtained by setting \verb{nls_v = "base", nls_weights=F}. This method weights heavily to early time points (higher concentration). Also, parameter constraints cannot be set with \code{nls()}, and so implausible estimates (negative for a, b) may be obtained.
\item \code{MSP-BM} The Multiple Sample Protocol (MSP), or Multiple Late Sample (MLS) estimate using 1-compartment analysis of Late time points only (>120 minutes post injection). Adjustment made using the Brochner-Mortenson correction equation to correct for unmeasured early time points.
\item \code{SS-Jacobbson} Single Sample estimate of GFR
}

For \code{output = "summary"} results returned includes the named variables:
\itemize{
\item \code{mgfr_method} Method used for GFR calculation (see above). Results should be similar.
\item \code{mgfr_2c} Measured GFR (mL/min)
\item \code{mgfr_2c_bsa} The measured GFR (mL/min/1.73 m2) indexed to body surface area (BSA calculated by DuBois equation)
\item \code{iohexol_m}   Iohexol mass (ucg) injected. Determined either by user-provided syringe weight or volume injected.
\item \code{iohexol_0}   Iohexol concentration at t=0 calculated by A+B (theoretical, not measured)
\item \code{iohexol_vd}  Iohexol Volume of distribution estimated by Dose/Iohexol\link{t0}
\item \code{ioh_auc}  Iohexol calculated AUC from t=0 to Infinity, estimated by A/a + B/b.
\item \code{A} Model parameter A
\item \code{a} Model parameter a, or \alpha
\item \code{B} Model parameter B
\item \code{b} Model parameter b, or \beta
\item \code{model_r2} Model Pseudo-R2, calculated as the linear regression R2 for predicted~observed values
\item \code{ssr} Sum of squared residuals
\item \code{sse} Sum of squared residuals for early time points
\item \code{ssl} Sum of squared residuals for late time points
\item \code{k10} Iohexol elimination rate constant from central compartment (1/min)
\item \code{k21} Iohexol transfer rate constant from peripheral to central compartment (1/min)
\item \code{k12} Iohexol transfer rate constant from central to peripheral compartment (1/min)
\item \code{n_early} Number of early time points measured and used in model
\item \code{n_late} Number of late time points measured and used in model
}

\code{gfr} returns a single \code{mgfr_2c} value, the measured GFR (mL/min).
\code{gfr_bsa} returns a single \code{mgfr_2c_bsa} value, the measured GFR
(mL/min/1.73 m2) indexed to body surface area (BSA calculated by DuBois
equation).

\code{fit} returns the non-linear regression model fit object.

\code{plot} returns a base-R plot of observed and predicted regression curve vs
time, and summary measures in the legend.}
}
\description{
This method uses multiple sample measurements of iohexol to determine GFR
(glomerular filtration rate), using a single compartment model of the late
slope and adjustment using the Brochner-Mortenson correction equation (.
Relevant Methods and recommendations are detailed in the
\href{https://pubmed.ncbi.nlm.nih.gov/39097002/}{European Kidney Function
Consortium statment paper}.
The function can be compared against an excel spreadsheet provided in the
supplemental material.
}
\examples{
library(tabletools)
library(dplyr)
# data from Ebert KI 2024, in online XLS file: https://pubmed.ncbi.nlm.nih.gov/39097002/
df1_dem <- data.frame(height=168,     # cm
                      weight=87,      # kg
                      ioh_inj_wt = 6.82594, # injected weight g; syringe wt pre-post
                      ioh_vol = 5.06, # mL
                      ioh_conc = 647)   # Omnipaque 300

dat_ebert

calculate_mgfr_msp(dat_ebert$time, dat_ebert$iohexol,
                   ioh_inj_vol = 5.06,
                   height = df1_dem$height, weight = df1_dem$weight)  
calculate_mgfr_msp(dat_ebert$time, dat_ebert$iohexol,
                   ioh_inj_wt = 6.82594, 
                   ioh_inj_vol=NULL, # make this NULL if weight given
                   height = df1_dem$height, weight = df1_dem$weight)  
                   
# methods designed to match the fuller for 2C
calculate_mgfr_msp(dat_ebert$time, dat_ebert$iohexol, height = 1.68, weight=87, ioh_inj_vol = 5.06, output = "summary")
calculate_mgfr_msp(dat_ebert$time, dat_ebert$iohexol, height = 1.68, weight=87, ioh_inj_vol = 5.06, output = "gfr")
calculate_mgfr_msp(dat_ebert$time, dat_ebert$iohexol, height = 1.68, weight=87, ioh_inj_vol = 5.06, output = "gfr_bsa")
calculate_mgfr_msp(dat_ebert$time, dat_ebert$iohexol, height = 1.68, weight=87, ioh_inj_vol = 5.06, output = "fit")
calculate_mgfr_msp(dat_ebert$time, dat_ebert$iohexol, height = 1.68, weight=87, ioh_inj_vol = 5.06, output = "plot", id="test")

# Example for multiple dataset analysis
df2_dem <- data.frame(id=1:5,
                      height= rnorm(5, 168, 10),     # cm
                     weight= rnorm(5, 87, 5),      # kg
                     ioh_inj_wt = rep(6.82594,5), # injected weight g; syringe wt pre-post
                     ioh_vol = rep(5.06,5)) # mL
dat2 <- data.frame(id=rep(1:5, each=6),
                  time = rep(c(160,180,200,220,232,240), 5),
                  iohexol_ucg_ml = rep(c(70,60,47,37,30,25), 5))

# merged with time series df in a column
df2_m <- merge(df2_dem, dat2, all.x = T) |> 
 dplyr::group_by(id) |> 
 tidyr::nest(dat = c(time, iohexol_ucg_ml))
library(purrr)
df2_m |> 
 mutate(mgfr = map(dat, ~calculate_mgfr_msp(.x$time, .x$iohexol_ucg_ml,
                                            ioh_inj_vol = ioh_vol,
                                            height = height, weight = weight))) |> 
 tidyr::unnest(mgfr)
}
