% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpeptide_params.R
\name{cpeptide_params}
\alias{cpeptide_params}
\title{C-peptide parameters estimates from population data}
\usage{
cpeptide_params(
  age,
  gender = NA,
  height,
  weight,
  category = "normal",
  weight_units = "kg",
  height_units = "m"
)
}
\arguments{
\item{age}{Age in years}

\item{gender}{Gender}

\item{height}{Height in meters}

\item{weight}{Weight in kg}

\item{category}{Category ("Normal", "Obese", "NIDDM")}

\item{weight_units}{weight units, if not in kg}

\item{height_units}{height units, if not in meters}
}
\value{
Dagta frame with Vd, halflife_short, halflife_long, fraction, bsa,
  k12, k21, k10
}
\description{
This function estimates C-peptide parameters from population data. Based on
the regression models from \href{https://pubmed.ncbi.nlm.nih.gov/1551497/}{Van Cauter et al}
}
\examples{
cpeptide_params(age=28.1, height = 1.744666, weight = 69.4, gender = "F", category = "normal")
cpeptide_params(age=28.1, height = 1.744666, weight = 69.4, gender = "M", category = "normal")
cpeptide_params(age=35.2, height = 1.678461, weight = 107.9, gender = "F", category = "obese")
cpeptide_params(age=35.2, height = 1.678461, weight = 107.9, gender = "M", category = "obese")
}
