% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabletools.R
\name{ftToPng}
\alias{ftToPng}
\title{Converts a FlexTable intpa png file}
\usage{
ftToPng(ft, zoom = 1, rotate = 0)
}
\arguments{
\item{ft}{The FlexTable object to convert}

\item{zoom}{Zoomm level}

\item{rotate}{Rotation 0 - 360}
}
\value{
Path of png file created
}
\description{
Converts a FlexTable intpa png file
}

