% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabletools.R
\name{htmlToPng}
\alias{htmlToPng}
\title{Converts any html into a png file}
\usage{
htmlToPng(html, zoom = 1, rotate = 0, vwidth = 992, vheight = 74,
  cliprect = NULL, selector = NULL)
}
\arguments{
\item{html}{HTML text to convert}

\item{zoom}{Zoomfactor}

\item{rotate}{Rotation 0 - 360}

\item{vwidth}{viewport width to set screenshot boundaries}

\item{vheight}{viewport height to set screenshot boundaries}

\item{cliprect}{use cliprect to set screenshot boundaries}

\item{selector}{use CSS selector toset screenshot boundaries}
}
\value{
A file name of the temporary png file created
}
\description{
Converts any html into a png file
}

