% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcorr_p_adjustment.R
\name{rcorr_padjust}
\alias{rcorr_padjust}
\title{Adjust p-value for an Hmisc::rcorr object}
\usage{
rcorr_padjust(x, method = "BH")
}
\arguments{
\item{x}{An `rcorr` class object created by the function `Hmisc::rcorr``}

\item{method}{Uses "BH" by default}
}
\value{
An rcorr array with the following matrices:
\itemize{
  \item correlation (`$r`). unchanged from the original
  \item n (`$n`). unchanged from the original
  \item P-value (`$P`). modified using the selected adjustment method
}
}
\description{
Modifies the `rcorr` p-values from the `Hmisc::rcorr()` function.
The result is an array with modified `$P` matrix.
uses `base::p.adjust`
}
\details{
Adjust Hmisc::rcorr p-values using `p.adjust`
}
\examples{
require(Hmisc)
require(Hmisc)
set.seed(2020) # arguably the worse seed ever
test_rcorr <- Hmisc::rcorr(matrix(runif(100), nrow = 10))
test_rcorr

# use p.adjust methods
rcorr_padjust(test_rcorr) # BH by default
rcorr_padjust(test_rcorr, method = "bonferroni") # BH by default
rcorr_padjust(test_rcorr, method = "holm") # BH by default
# SUMMARY TABLE OF RESULTS
# one unadjusted result has a low P-value (p=0.008)
method = c("none", "BH", "fdr", "holm", "hochberg", "hommel", "bonferroni", "BY")
data.frame(p.adjust.method = method,
           p.adjust = map_dbl(method, ~rcorr_padjust(test_rcorr, method = .x)$P[4,5]),
          txtPval = htmlTable::txtPval(map_dbl(method, ~rcorr_padjust(test_rcorr, method = .x)$P[4,5])))
}
\seealso{
\code{\link[Hmisc]{rcorr}}, \code{\link[heatmaply]{heatmaply}}
}
