% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_grouped_summary.R
\name{table_grouped_summary}
\alias{table_grouped_summary}
\title{Grouped Summary Table with optional html output}
\usage{
table_grouped_summary(
  df,
  var_names,
  grouping_variable,
  html_output = FALSE,
  total = TRUE
)
}
\arguments{
\item{df}{a dataframe}

\item{var_names}{a character vector containing bare variable names}

\item{grouping_variable}{a character string containing the grouping variable name}

\item{html_output}{should the result be returned as an htmlTable (if `html_table`` = TRUE) or as a dataframe (default)}

\item{total}{should an overal total column be included (default) or omitted}
}
\value{
A text table with grouped summary
}
\description{
Grouped Summary Table with optional html output
}
\examples{
library(tidyverse)
          by = list(cyl = mtcars[, "cyl"]),
          FUN = function(x) sprintf("\%.2f &pm \%.2f", mean(x), sd(x))) \%>\% 
  t() \%>\% 
  as.data.frame() \%>\% 
  rownames_to_column() \%>\% 
  {names(.) <- unlist(.[1, ]); .} \%>\% 
  .[-1, ] 

# output HTML format
mtcars \%>\% 
  table_grouped_summary(., 
                        var_names =  c("disp", "hp", "drat", "mpg", "wt"),
                        grouping_variable = "cyl",
                        html_output = TRUE,
                        total = TRUE)

# output text format
  mtcars \%>\% 
  table_grouped_summary(., 
                        var_names =  c("disp", "hp", "drat", "mpg", "wt"),
                        grouping_variable = "cyl",
                        html_output = FALSE,
                        total = TRUE)

# without overall total
mtcars \%>\% 
  table_grouped_summary(., 
                        var_names =  c("disp", "hp", "drat", "mpg", "wt"),
                        grouping_variable = "cyl",
                        html_output = FALSE,
                        total = FALSE)
}
\seealso{
\code{\link[base]{sprintf}}, \code{\link[base]{mean}}, \code{\link[stats]{sd}}, \code{\link[htmlTable]{htmlTable}}
}
