% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt_mean_range.R
\name{txt_mean_range}
\alias{txt_mean_range}
\title{Summary text functions for html tables}
\usage{
txt_mean_range(var, na.rm = TRUE, digits = 1, unicode = T, ...)
}
\arguments{
\item{var}{A numeric vector or variable name within a dataframe}

\item{na.rm}{a logical value indicating whether NA values should be stripped}

\item{digits}{number of decimal places (limited to 1-3)}
}
\value{
A character vector with mean (range), suitable for printing in an html table
}
\description{
Summary text functions for html tables
}
\examples{
library(dplyr)
txt_mean_range(mtcars$mpg)
mtcars \%>\%
  group_by(cyl) \%>\%
  summarise(mean_mpg = txt_mean_range(mpg))
}
\seealso{
\code{\link[base]{sprintf}}, \code{\link[base]{mean}}, \code{\link[base]{min}}, \code{\link[base]{max}}
}
