% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt_mean_sd.R
\name{txt_mean_sd}
\alias{txt_mean_sd}
\title{Summary text functions for html tables}
\usage{
txt_mean_sd(var, na.rm = TRUE, digits = 2)
}
\arguments{
\item{var}{A numeric vector or variable name within a dataframe}

\item{na.rm}{a logical value indicating whether NA values should be stripped}

\item{digits}{number of decimal places (limited to 1-3)}
}
\value{
A character vector with mean and sd, suitable for printing in an html table
}
\description{
Summary text functions for html tables
}
\examples{
library(dplyr)
txt_mean_sd(1:10)
txt_mean_sd(1:10, digits = 1)
txt_mean_sd(1:10, digits = 3)
}
\seealso{
\code{\link[base]{sprintf}}, \code{\link[base]{mean}}, \code{\link[base]{sd}}
}
